refactor(runtime): split utils.rs into cohesive submodules (§7-8)
Split the 3954-line runtime/utils.rs (over the 500-line limit) into a
facade plus twelve themed submodules under src/runtime/utils/, all under
500 lines. utils.rs is entirely free functions addressed as
crate::runtime::utils::<name>; to keep every caller path working it now
glob-re-exports each submodule (pub(crate) use sets::*; etc.), so no
external caller changes.

- utils.rs (314)               - facade: use block, misc small fns,
                                 make_order/version_cmp_parts, re-exports
- coerce_containers.rs (435)   - coerce_to_hash/array/str, build_hash_from_items
- type_constraints.rs (479)    - is_known_type_constraint / compound
- list.rs (393)                - char_idx_to_byte, value_to_list, etc.
- gist.rs (323)                - gist_value, match_gist, setbagmix_gist
- compare.rs (322)             - to_float_value, compare_values, to_int
- set_ops.rs (301)             - set diff/intersect/sym-diff
- set_coerce.rs (278)          - coerce_to_set/quanthash + bag/mix maps
- shaped.rs (264)              - shaped-array helpers + values_identical
- type_misc.rs (247)           - value_type_name, reduction_identity
- radix_numeric.rs (231)       - parse_radix_number_body, coerce_to_numeric
- errors.rs (221)              - type-check error builders, merge_junction
- rat.rs (181)                 - Rat parts helpers + numeric coercion errors

Pure relocation. Private helpers stay with their callers; three
(set_type_level/to_mix_map/to_bag_map) bumped to pub(crate) for
cross-module use. runtime/mod.rs is untouched (submodules are declared
inside the facade). The fn multiset is byte-identical. No behavior change.
