+++ src/runtime/utils.rs
@@ -212,270 +212,6 @@ pub(crate) fn hash_typed_key(hash: &Value, str_key: &str) -> Value {
     }
     Value::str(str_key.to_string())
 }
-
-/// Check if an array is a shaped (multidimensional) array.
-/// A shaped array is one explicitly created as multidimensional via `:shape`.
-pub(crate) fn is_shaped_array(value: &Value) -> bool {
-    if let Value::Array(_, kind) = value
-        && *kind == ArrayKind::Shaped
-    {
-        return true;
-    }
-    shaped_array_shape(value).is_some()
-}
-
-pub(crate) fn shaped_array_shape(value: &Value) -> Option<Vec<usize>> {
-    let Value::Array(items, kind) = value else {
-        return None;
-    };
-    // Only arrays explicitly created as shaped can be shaped
-    if *kind != ArrayKind::Shaped {
-        return None;
-    }
-
-    fn shape_matches_structure(value: &Value, shape: &[usize]) -> bool {
-        if shape.is_empty() {
-            return false;
-        }
-        let Value::Array(items, ..) = value else {
-            return false;
-        };
-        if items.len() != shape[0] {
-            return false;
-        }
-        if shape.len() == 1 {
-            return items.iter().all(|v| !matches!(v, Value::Array(..)));
-        }
-        items
-            .iter()
-            .all(|child| shape_matches_structure(child, &shape[1..]))
-    }
-
-    if items.is_empty() {
-        return None;
-    }
-
-    // Infer this array's shape from its children's embedded shapes.
-    fn infer_shape_from_array(items: &[Value]) -> Option<Vec<usize>> {
-        let first = items.first()?;
-        let Value::Array(first_items, ..) = first else {
-            return None;
-        };
-        let first_shape = first_items.shape.as_ref()?;
-        if !items.iter().all(|child| {
-            if let Value::Array(child_items, ..) = child {
-                child_items.shape.as_ref() == Some(first_shape)
-            } else {
-                false
-            }
-        }) {
-            return None;
-        }
-        let mut shape = Vec::with_capacity(1 + first_shape.len());
-        shape.push(items.len());
-        shape.extend_from_slice(first_shape);
-        Some(shape)
-    }
-
-    // Prefer the shape embedded on this array's `ArrayData`, validated against
-    // the current element structure (a stale cached shape after restructuring
-    // is rejected and re-inferred).
-    if let Some(cached_shape) = &items.shape
-        && shape_matches_structure(value, cached_shape)
-    {
-        return Some(cached_shape.clone());
-    }
-    let inferred_shape = infer_shape_from_array(items.as_ref())?;
-    if !shape_matches_structure(value, &inferred_shape) {
-        return None;
-    }
-    mark_shaped_array_items(items, Some(&inferred_shape));
-    Some(inferred_shape)
-}
-
-pub(crate) fn mark_shaped_array(value: &Value, shape: Option<&[usize]>) {
-    let Value::Array(items, ..) = value else {
-        return;
-    };
-    mark_shaped_array_items(items, shape);
-}
-
-pub(crate) fn mark_shaped_array_items(
-    items: &Arc<crate::value::ArrayData>,
-    shape: Option<&[usize]>,
-) {
-    let Some(shape) = shape else {
-        return;
-    };
-    if items.shape.as_deref() == Some(shape) {
-        return;
-    }
-    // The shape is metadata about this one logical array, shared by every holder
-    // of the `Arc` — matching the prior pointer-keyed side-table semantics (any
-    // holder of the same pointer saw the shape). Interior mutation preserves
-    // "mark after the array Value is already placed" without a write-back through
-    // the caller.
-    // SAFETY: aliased in-place mutation of a shared container; see
-    // `arc_contents_mut`. No borrow into the array is live across this write.
-    unsafe {
-        crate::value::arc_contents_mut(items).shape = Some(shape.to_vec());
-    }
-}
-
-/// Collect all leaf values from a shaped (multidimensional) array.
-pub(crate) fn shaped_array_leaves(value: &Value) -> Vec<Value> {
-    let mut leaves = Vec::new();
-    collect_leaves(value, &mut leaves);
-    leaves
-}
-
-fn collect_leaves(value: &Value, out: &mut Vec<Value>) {
-    if let Value::Array(items, ..) = value {
-        if items.iter().any(|v| matches!(v, Value::Array(..))) {
-            for item in items.iter() {
-                collect_leaves(item, out);
-            }
-        } else {
-            out.extend(items.iter().cloned());
-        }
-    } else {
-        out.push(value.clone());
-    }
-}
-
-/// Collect all (index-tuple, leaf-value) pairs from a shaped array.
-pub(crate) fn shaped_array_indexed_leaves(value: &Value) -> Vec<(Vec<i64>, Value)> {
-    let mut result = Vec::new();
-    let mut indices = Vec::new();
-    collect_indexed_leaves(value, &mut indices, &mut result);
-    result
-}
-
-fn collect_indexed_leaves(value: &Value, indices: &mut Vec<i64>, out: &mut Vec<(Vec<i64>, Value)>) {
-    if let Value::Array(items, ..) = value {
-        if items.iter().any(|v| matches!(v, Value::Array(..))) {
-            for (i, item) in items.iter().enumerate() {
-                indices.push(i as i64);
-                collect_indexed_leaves(item, indices, out);
-                indices.pop();
-            }
-        } else {
-            for (i, item) in items.iter().enumerate() {
-                let mut idx = indices.clone();
-                idx.push(i as i64);
-                out.push((idx, item.clone()));
-            }
-        }
-    }
-}
-
-pub(crate) fn values_identical(left: &Value, right: &Value) -> bool {
-    match (left, right) {
-        (Value::Package(name), Value::Int(0)) | (Value::Int(0), Value::Package(name))
-            if name.resolve() == "int" =>
-        {
-            true
-        }
-        (Value::Array(a, _), Value::Array(b, _)) => std::sync::Arc::ptr_eq(a, b),
-        (Value::Seq(a), Value::Seq(b)) => std::sync::Arc::ptr_eq(a, b),
-        (Value::Slip(a), Value::Slip(b)) => {
-            // Empty is a singleton semantic value in Raku even when represented
-            // by distinct empty Slip allocations.
-            (a.is_empty() && b.is_empty()) || std::sync::Arc::ptr_eq(a, b)
-        }
-        (Value::LazyList(a), Value::LazyList(b)) => std::sync::Arc::ptr_eq(a, b),
-        (Value::Hash(a), Value::Hash(b)) => std::sync::Arc::ptr_eq(a, b),
-        (Value::Sub(a), Value::Sub(b)) => {
-            if std::sync::Arc::ptr_eq(a, b) {
-                return true;
-            }
-            // Named subs with the same package and name are identical
-            // (e.g. &foo === &EXPORT::ALL::foo when both resolve to the same definition)
-            let a_name = a.name.resolve();
-            let b_name = b.name.resolve();
-            !a_name.is_empty() && a_name == b_name && a.package == b.package
-        }
-        (Value::WeakSub(a), Value::WeakSub(b)) => a.ptr_eq(b),
-        (Value::Mixin(a_inner, a_mix), Value::Mixin(b_inner, b_mix)) => {
-            a_inner.eqv(b_inner) && a_mix == b_mix
-        }
-        (Value::Mixin(_, _), _) | (_, Value::Mixin(_, _)) => false,
-        (
-            Value::Instance {
-                class_name: a_class,
-                id: a_id,
-                attributes: a_attrs,
-                ..
-            },
-            Value::Instance {
-                class_name: b_class,
-                id: b_id,
-                attributes: b_attrs,
-                ..
-            },
-        ) => {
-            let a_name = a_class.resolve();
-            let b_name = b_class.resolve();
-            if a_name == b_name
-                && (a_name == "Stash" || a_name == "Supply" || a_name == "IO::Special")
-            {
-                left.eqv(right)
-            } else if a_name == b_name && (a_name == "ObjAt" || a_name == "ValueObjAt") {
-                // ObjAt/ValueObjAt instances are === when their WHICH content matches
-                let a_val = a_attrs.as_map().get("WHICH").map(|v| v.to_string_value());
-                let b_val = b_attrs.as_map().get("WHICH").map(|v| v.to_string_value());
-                a_val == b_val
-            } else {
-                a_id == b_id
-            }
-        }
-        // Junction identity: each junction object is unique
-        (Value::Junction { values: a_vals, .. }, Value::Junction { values: b_vals, .. }) => {
-            std::sync::Arc::ptr_eq(a_vals, b_vals)
-        }
-        // Capture identity is NOT structural: a Capture's `.WHICH` keeps the
-        // *container* identity of each captured element, so `\($a) === \($b)`
-        // is False even when `$a` and `$b` hold equal values, while
-        // `\(42) === \(42)` is True (both literal-value elements). This is
-        // unlike `eqv`, which deconts and compares structurally.
-        (
-            Value::Capture {
-                positional: ap,
-                named: an,
-            },
-            Value::Capture {
-                positional: bp,
-                named: bn,
-            },
-        ) => {
-            ap.len() == bp.len()
-                && an.len() == bn.len()
-                && ap
-                    .iter()
-                    .zip(bp.iter())
-                    .all(|(x, y)| capture_elem_identical(x, y))
-                && an
-                    .iter()
-                    .all(|(k, v)| bn.get(k).is_some_and(|bv| capture_elem_identical(v, bv)))
-        }
-        _ => left.eqv(right),
-    }
-}
-
-/// Identity comparison for a single Capture element. Unlike a top-level
-/// `===` (which deconts a bound scalar to its value), a Capture retains the
-/// container identity of each element: two `ContainerRef` cells are identical
-/// only when they are the same `Arc` (i.e. bound to the same container), and a
-/// container can never be identical to a plain value. Non-container elements
-/// fall back to the normal value identity rules.
-fn capture_elem_identical(a: &Value, b: &Value) -> bool {
-    match (a, b) {
-        (Value::ContainerRef(x), Value::ContainerRef(y)) => std::sync::Arc::ptr_eq(x, y),
-        (Value::ContainerRef(_), _) | (_, Value::ContainerRef(_)) => false,
-        _ => values_identical(a, b),
-    }
-}
-
 pub(crate) fn make_order(ord: std::cmp::Ordering) -> Value {
     match ord {
         std::cmp::Ordering::Less => Value::Enum {
@@ -547,3408 +283,32 @@ pub(crate) fn version_cmp_parts(
     std::cmp::Ordering::Equal
 }
 
-pub(crate) fn coerce_to_hash(value: Value) -> Value {
-    let mix_weight_value = crate::value::mix_weight_to_value;
-    let value = value.into_descalarized();
-    match value {
-        Value::Hash(_) => value,
-        Value::Array(items, ..) => {
-            // Flatten nested Hashes into pairs before building the hash.
-            // This handles `%h = %h1, %h2` where each hash should be merged.
-            // Itemized arrays ($[...]) are NOT flattened — they are treated
-            // as opaque items, matching Raku's Scalar-container semantics.
-            let mut flat: Vec<Value> = Vec::with_capacity(items.len());
-            for item in items.iter() {
-                if let Value::Hash(h) = item {
-                    for (k, v) in h.iter() {
-                        flat.push(Value::Pair(k.clone(), Box::new(v.clone())));
-                    }
-                } else {
-                    flat.push(item.clone());
-                }
-            }
-            let mut map = HashMap::new();
-            let mut original_keys: HashMap<String, Value> = HashMap::new();
-            let mut i = 0;
-            while i < flat.len() {
-                if let Value::Pair(k, v) = &flat[i] {
-                    // A Pair value built by `key => $var` is a write-through
-                    // `ContainerRef`; storing into a Hash decontainerizes (copies
-                    // the value), matching Raku (`%h = k => $v; $v = 2` leaves
-                    // `%h<k>` unchanged).
-                    map.insert(k.clone(), v.deref_container());
-                    i += 1;
-                } else if let Value::ValuePair(k, v) = &flat[i] {
-                    let str_key = k.to_string_value();
-                    if !matches!(k.as_ref(), Value::Str(_)) {
-                        original_keys.insert(str_key.clone(), *k.clone());
-                    }
-                    map.insert(str_key, v.deref_container());
-                    i += 1;
-                } else {
-                    let key_val = &flat[i];
-                    let str_key = key_val.to_string_value();
-                    if !matches!(key_val, Value::Str(_)) {
-                        original_keys.insert(str_key.clone(), key_val.clone());
-                    }
-                    let val = if i + 1 < flat.len() {
-                        flat[i + 1].clone()
-                    } else {
-                        Value::Nil
-                    };
-                    map.insert(str_key, val);
-                    i += 2;
-                }
-            }
-            set_hash_original_keys(Value::hash(map), original_keys)
-        }
-        Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) | Value::Slip(items) => {
-            let mut map = HashMap::new();
-            let mut i = 0;
-            while i < items.len() {
-                if let Value::Pair(k, v) = &items[i] {
-                    map.insert(k.clone(), v.deref_container());
-                    i += 1;
-                } else if let Value::ValuePair(k, v) = &items[i] {
-                    map.insert(k.to_string_value(), v.deref_container());
-                    i += 1;
-                } else {
-                    let key = items[i].to_string_value();
-                    let val = if i + 1 < items.len() {
-                        items[i + 1].clone()
-                    } else {
-                        Value::Nil
-                    };
-                    map.insert(key, val);
-                    i += 2;
-                }
-            }
-            Value::hash(map)
-        }
-        Value::Pair(k, v) => {
-            let mut map = HashMap::new();
-            map.insert(k, v.deref_container());
-            Value::hash(map)
-        }
-        Value::ValuePair(k, v) => {
-            let mut map = HashMap::new();
-            map.insert(k.to_string_value(), v.deref_container());
-            Value::hash(map)
-        }
-        Value::Set(items, _) => {
-            let mut map = HashMap::new();
-            let mut original_keys: HashMap<String, Value> = HashMap::new();
-            let mut has_typed = false;
-            for key in items.iter() {
-                map.insert(key.clone(), Value::Bool(true));
-                let typed = items.typed_key(key);
-                if !matches!(&typed, Value::Str(sv) if sv.as_ref() == key) {
-                    has_typed = true;
-                    original_keys.insert(key.clone(), typed);
-                }
-            }
-            let mut result = Value::hash(map);
-            if has_typed {
-                original_keys.insert("__mutsu_setty_origin".to_string(), Value::Bool(true));
-                result = set_hash_original_keys(result, original_keys);
-            }
-            result
-        }
-        Value::Bag(items, _) => {
-            let mut map = HashMap::new();
-            let mut original_keys: HashMap<String, Value> = HashMap::new();
-            let mut has_typed = false;
-            for (key, count) in items.iter() {
-                map.insert(key.clone(), Value::from_bigint(count.clone()));
-                let typed = items.typed_key(key);
-                if !matches!(&typed, Value::Str(sv) if sv.as_ref() == key) {
-                    has_typed = true;
-                    original_keys.insert(key.clone(), typed);
-                }
-            }
-            let mut result = Value::hash(map);
-            if has_typed {
-                original_keys.insert("__mutsu_setty_origin".to_string(), Value::Bool(true));
-                result = set_hash_original_keys(result, original_keys);
-            }
-            result
-        }
-        Value::Mix(items, _) => {
-            let mut map = HashMap::new();
-            let mut original_keys: HashMap<String, Value> = HashMap::new();
-            let mut has_typed = false;
-            for (key, weight) in items.iter() {
-                map.insert(key.clone(), mix_weight_value(*weight));
-                let typed = items.typed_key(key);
-                if !matches!(&typed, Value::Str(sv) if sv.as_ref() == key) {
-                    has_typed = true;
-                    original_keys.insert(key.clone(), typed);
-                }
-            }
-            let mut result = Value::hash(map);
-            if has_typed {
-                original_keys.insert("__mutsu_setty_origin".to_string(), Value::Bool(true));
-                result = set_hash_original_keys(result, original_keys);
-            }
-            result
-        }
-        Value::Range(a, b) => {
-            let items: Vec<Value> = (a..=b).map(Value::Int).collect();
-            coerce_to_hash(Value::Array(
-                crate::value::Value::array_arc(items),
-                ArrayKind::List,
-            ))
-        }
-        Value::RangeExcl(a, b) => {
-            let items: Vec<Value> = (a..b).map(Value::Int).collect();
-            coerce_to_hash(Value::Array(
-                crate::value::Value::array_arc(items),
-                ArrayKind::List,
-            ))
-        }
-        Value::RangeExclStart(a, b) => {
-            let items: Vec<Value> = (a + 1..=b).map(Value::Int).collect();
-            coerce_to_hash(Value::Array(
-                crate::value::Value::array_arc(items),
-                ArrayKind::List,
-            ))
-        }
-        Value::RangeExclBoth(a, b) => {
-            let items: Vec<Value> = (a + 1..b).map(Value::Int).collect();
-            coerce_to_hash(Value::Array(
-                crate::value::Value::array_arc(items),
-                ArrayKind::List,
-            ))
-        }
-        Value::Nil => Value::hash(HashMap::new()),
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Match" => {
-            // %($/) returns the named captures hash
-            if let Some(named) = attributes.as_map().get("named") {
-                named.clone()
-            } else {
-                Value::hash(HashMap::new())
-            }
-        }
-        _ => {
-            let mut map = HashMap::new();
-            map.insert(value.to_string_value(), Value::Nil);
-            Value::hash(map)
-        }
-    }
-}
+mod coerce_containers;
+mod compare;
+mod errors;
+mod gist;
+mod list;
+mod radix_numeric;
+mod rat;
+mod set_coerce;
+mod set_ops;
+mod shaped;
+mod type_constraints;
+mod type_misc;
 
-pub(crate) fn build_hash_from_items(items: Vec<Value>) -> Result<Value, RuntimeError> {
-    let total_items = items.len();
-    let last_item = items
-        .last()
-        .map(Value::to_string_value)
-        .unwrap_or_else(|| "Nil".to_string());
-    let mut map = HashMap::new();
-    let mut original_keys: HashMap<String, Value> = HashMap::new();
-    let mut iter = items.into_iter();
-    while let Some(item) = iter.next() {
-        match item {
-            Value::Pair(key, boxed_val) => {
-                map.insert(key, *boxed_val);
-            }
-            // A bare (non-itemized) hash in list context flattens into its
-            // key=>value pairs (`%m = (%h,)` / `%(%h,)`). A hash sourced from a
-            // `$` scalar carries `HashData.itemized` (set by `itemize_value`) and
-            // stays an opaque single element — matching Raku, where `%m =
-            // ($hashitem,)` dies "Odd number". Keys are the source hash's string
-            // keys: flattening an object hash into a plain Hash/Map stringifies
-            // them (the target re-tags via its own key-type metadata).
-            Value::Hash(ref h) if !h.itemized => {
-                for (k, v) in h.iter() {
-                    map.insert(k.clone(), v.clone());
-                }
-            }
-            Value::ValuePair(key, boxed_val) => {
-                let str_key = Value::hash_key_encode(&key);
-                if !matches!(key.as_ref(), Value::Str(_)) {
-                    original_keys.insert(str_key.clone(), *key);
-                }
-                map.insert(str_key, *boxed_val);
-            }
-            other => {
-                let Some(value) = iter.next() else {
-                    let message = format!(
-                        "Odd number of elements found where hash initializer expected: found {total_items} element(s); last element seen: {last_item}"
-                    );
-                    let mut attrs = std::collections::HashMap::new();
-                    attrs.insert(
-                        "message".to_string(),
-                        crate::value::Value::str(message.clone()),
-                    );
-                    let ex = crate::value::Value::make_instance(
-                        crate::symbol::Symbol::intern("X::Hash::Store::OddNumber"),
-                        attrs,
-                    );
-                    let mut err = RuntimeError::new(message);
-                    err.exception = Some(Box::new(ex));
-                    return Err(err);
-                };
-                let str_key = Value::hash_key_encode(&other);
-                if !matches!(&other, Value::Str(_)) {
-                    original_keys.insert(str_key.clone(), other);
-                }
-                map.insert(str_key, value);
-            }
-        }
-    }
-    Ok(set_hash_original_keys(Value::hash(map), original_keys))
-}
-
-/// Maximum number of elements when expanding an infinite range into an Array.
-/// TODO: Properly implement lazy arrays that reify elements on demand.
-const MAX_ARRAY_EXPAND: i64 = 100_000;
-
-pub(crate) fn coerce_to_array(value: Value) -> Value {
-    fn metadata_shape_for_items(items: &Arc<crate::value::ArrayData>) -> Option<Vec<usize>> {
-        items.shape.clone()
-    }
-
-    match value {
-        Value::Array(items, kind) => {
-            // Assigning an array to an `@` variable snapshots element VALUES
-            // (Raku `=` semantics). A `:=`-bound element is a shared
-            // `ContainerRef` cell (Phase 2); decontainerize it on copy so a
-            // later write through the bound source does not leak into the copy.
-            // Only rebuild when a cell is actually present (common path keeps
-            // sharing the Arc, so there is no per-assignment cost).
-            let items = if items.iter().any(|v| matches!(v, Value::ContainerRef(_))) {
-                Arc::new(items.iter().map(|v| v.deref_container()).collect())
-            } else {
-                items
-            };
-            if kind.is_itemized() {
-                // Itemized arrays (from `$` scalar containers) are treated as
-                // a single item when assigned to an `@` variable.
-                Value::real_array(vec![Value::Array(items, kind)])
-            } else if kind == ArrayKind::Shaped {
-                Value::Array(items, kind)
-            } else if let Some(shape) = metadata_shape_for_items(&items) {
-                let value = Value::Array(items, ArrayKind::Shaped);
-                mark_shaped_array(&value, Some(&shape));
-                value
-            } else {
-                Value::Array(items, ArrayKind::Array)
-            }
-        }
-        Value::Nil => Value::real_array(vec![Value::Package(crate::symbol::Symbol::intern("Any"))]),
-        Value::Range(a, b) => {
-            if b == i64::MAX {
-                // Infinite range — mark as lazy
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::Array(
-                    Arc::new((a..=end).map(Value::Int).collect()),
-                    ArrayKind::Lazy,
-                )
-            } else {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::real_array((a..=end).map(Value::Int).collect())
-            }
-        }
-        Value::RangeExcl(a, b) => {
-            if b == i64::MAX {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::Array(
-                    Arc::new((a..end).map(Value::Int).collect()),
-                    ArrayKind::Lazy,
-                )
-            } else {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::real_array((a..end).map(Value::Int).collect())
-            }
-        }
-        Value::RangeExclStart(a, b) => {
-            if b == i64::MAX {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::Array(
-                    Arc::new((a + 1..=end).map(Value::Int).collect()),
-                    ArrayKind::Lazy,
-                )
-            } else {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::real_array((a + 1..=end).map(Value::Int).collect())
-            }
-        }
-        Value::RangeExclBoth(a, b) => {
-            if b == i64::MAX {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::Array(
-                    Arc::new((a + 1..end).map(Value::Int).collect()),
-                    ArrayKind::Lazy,
-                )
-            } else {
-                let end = b.min(a.saturating_add(MAX_ARRAY_EXPAND));
-                Value::real_array((a + 1..end).map(Value::Int).collect())
-            }
-        }
-        Value::GenericRange {
-            ref start, ref end, ..
-        } if matches!(start.as_ref(), Value::Str(_)) && matches!(end.as_ref(), Value::Str(_)) => {
-            Value::real_array(value_to_list(&value))
-        }
-        Value::GenericRange { ref end, .. } => {
-            let end_f = end.to_f64();
-            if end_f.is_infinite() && end_f.is_sign_positive() {
-                Value::Array(
-                    Arc::new(crate::value::ArrayData::new(value_to_list(&value))),
-                    ArrayKind::Lazy,
-                )
-            } else {
-                Value::real_array(value_to_list(&value))
-            }
-        }
-        Value::Slip(items) | Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) => {
-            Value::Array(
-                crate::value::Value::array_arc(items.to_vec()),
-                ArrayKind::Array,
-            )
-        }
-        Value::LazyList(_) => value,
-        // A bare Hash assigned to an @-var flattens into its pairs
-        // (`my @a = %h`). An *itemized* hash (`my $h = %(...); my @a = $h`)
-        // stays a single element — `ItemizeVar` now tags it with
-        // `HashData.itemized` rather than the old `Scalar(Hash)` wrapper, so it
-        // must be handled here too (the `Scalar(inner)` arm below covers the
-        // Set/Bag/Mix itemized forms, which still use the wrapper).
-        Value::Hash(ref map) if !map.itemized => {
-            let pairs: Vec<Value> = map
-                .iter()
-                .map(|(k, v)| map.typed_pair(k, v.clone()))
-                .collect();
-            Value::real_array(pairs)
-        }
-        Value::Hash(map) => Value::real_array(vec![Value::Hash(map)]),
-        // A scalar holding a Hash/Set/Bag/Mix (itemized by `ItemizeVar` as
-        // `Scalar(container)`) stays a single element, but unwrap the Scalar so
-        // `@a[0]` is the bare container (preserving its `.gist`/`.raku`/type),
-        // rather than an opaque `$(...)`-wrapped value.
-        Value::Scalar(inner)
-            if matches!(
-                inner.as_ref(),
-                Value::Hash(_) | Value::Set(..) | Value::Bag(..) | Value::Mix(..)
-            ) =>
-        {
-            Value::real_array(vec![*inner])
-        }
-        // Set/Bag/Mix assigned to an @-var flatten into their `key => weight`
-        // pairs in list context, exactly like a Hash (Raku: `my @a = set(1,2,3)`
-        // yields three `* => True` pairs, so `@a.elems == 3`). This mirrors
-        // `value_to_list`. (Note: an array *literal* `[set(...)]` does NOT
-        // flatten — that path is handled separately in `exec_make_array_op`.)
-        Value::Set(ref items, _) => Value::real_array(
-            items
-                .iter()
-                .map(|s| Value::Pair(s.clone(), Box::new(Value::Bool(true))))
-                .collect(),
-        ),
-        Value::Bag(ref items, _) => Value::real_array(
-            items
-                .iter()
-                .map(|(k, v)| Value::Pair(k.clone(), Box::new(Value::from_bigint(v.clone()))))
-                .collect(),
-        ),
-        Value::Mix(ref items, _) => Value::real_array(
-            items
-                .iter()
-                .map(|(k, v)| {
-                    Value::Pair(k.clone(), Box::new(crate::value::mix_weight_to_value(*v)))
-                })
-                .collect(),
-        ),
-        // A WalkList assigned to an `@` variable flattens to its candidate
-        // closures, so `my @cands = $x.WALK(...)` yields the per-level candidates.
-        Value::Instance {
-            ref class_name,
-            ref attributes,
-            ..
-        } if class_name.resolve() == "WalkList" => match walk_list_candidates(attributes) {
-            Some(cands) => Value::real_array(cands),
-            None => Value::real_array(vec![value.clone()]),
-        },
-        other => Value::real_array(vec![other]),
-    }
-}
-
-pub(crate) fn coerce_to_str(value: &Value) -> String {
-    value.to_str_context()
-}
-
-/// Render the `.gist` form of a Set/Bag/Mix (and their mutable `*Hash`
-/// variants): `Set(a b c)`, `Bag(a b(2))`, `Mix(a(1.5) b)`. Keys are sorted
-/// for deterministic output. Returns `None` for any other value. Shared by
-/// `gist_value` (the fast say/gist path) and the `.gist` method dispatch so
-/// both render identically.
-pub(crate) fn setbagmix_gist(value: &Value) -> Option<String> {
-    match value {
-        Value::Set(s, mutable) => {
-            let type_name = if *mutable { "SetHash" } else { "Set" };
-            let mut keys: Vec<&String> = s.iter().collect();
-            keys.sort();
-            // Render each element via its original type's gist (so a Pair element
-            // is `a => 1`, not the internal `a\t1` string key); a plain Str/Int
-            // element gists bare, identical to the raw key.
-            let inner = keys
-                .iter()
-                .map(|k| gist_value(&s.typed_key(k)))
-                .collect::<Vec<_>>()
-                .join(" ");
-            Some(format!("{}({})", type_name, inner))
-        }
-        Value::Bag(b, mutable) => {
-            let type_name = if *mutable { "BagHash" } else { "Bag" };
-            let mut keys: Vec<(&String, &BigInt)> = b.iter().collect();
-            keys.sort_by_key(|(k, _)| (*k).clone());
-            let inner = keys
-                .iter()
-                .map(|(k, v)| {
-                    let key = gist_value(&b.typed_key(k));
-                    if **v == BigInt::from(1) {
-                        key
-                    } else {
-                        format!("{}({})", key, v)
-                    }
-                })
-                .collect::<Vec<_>>()
-                .join(" ");
-            Some(format!("{}({})", type_name, inner))
-        }
-        Value::Mix(m, mutable) => {
-            let type_name = if *mutable { "MixHash" } else { "Mix" };
-            let mut keys: Vec<(&String, &f64)> = m.iter().collect();
-            keys.sort_by_key(|(k, _)| (*k).clone());
-            let inner = keys
-                .iter()
-                .map(|(k, v)| {
-                    let key = gist_value(&m.typed_key(k));
-                    if (**v - 1.0).abs() < f64::EPSILON {
-                        key
-                    } else if v.fract() == 0.0 {
-                        format!("{}({})", key, **v as i64)
-                    } else {
-                        format!("{}({})", key, v)
-                    }
-                })
-                .collect::<Vec<_>>()
-                .join(" ");
-            Some(format!("{}({})", type_name, inner))
-        }
-        _ => None,
-    }
-}
-
-pub(crate) fn gist_value(value: &Value) -> String {
-    // Cycle detection for recursive data structures (shared hash/array Arcs).
-    thread_local! {
-        static SEEN_PTRS: std::cell::RefCell<Vec<usize>> = const { std::cell::RefCell::new(Vec::new()) };
-    }
-    fn check_and_push(ptrs: &std::cell::RefCell<Vec<usize>>, ptr: usize) -> bool {
-        let mut s = ptrs.borrow_mut();
-        if s.contains(&ptr) {
-            return true; // cycle detected
-        }
-        s.push(ptr);
-        false
-    }
-    fn pop_ptr(ptrs: &std::cell::RefCell<Vec<usize>>, ptr: usize) {
-        let mut s = ptrs.borrow_mut();
-        if let Some(pos) = s.iter().rposition(|p| *p == ptr) {
-            s.remove(pos);
-        }
-    }
-    match value {
-        // A `:=`-bound element holds a `ContainerRef` cell; render the held
-        // value so a bound element gists like a plain one (Phase 5 leak).
-        Value::ContainerRef(cell) => gist_value(&cell.lock().unwrap()),
-        Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _) => {
-            // Rat.gist is identical to Rat.Str in Raku
-            value.to_string_value()
-        }
-        Value::Array(_, crate::value::ArrayKind::Lazy) => {
-            // A lazy (infinite-backed) array renders a bounded placeholder
-            // rather than materializing its capped backing (Rakudo: `[...]`).
-            "[...]".to_string()
-        }
-        Value::LazyList(ll) if ll.is_genuinely_lazy() => {
-            // A genuinely-lazy list renders raku's placeholder without forcing:
-            // `[...]` held in `@` array context, `(...)` for a bare Seq.
-            crate::value::lazy_list_placeholder("gist", ll.in_array_context())
-        }
-        Value::Array(items, kind) => {
-            let ptr = std::sync::Arc::as_ptr(items) as usize;
-            let is_cycle = SEEN_PTRS.with(|seen| check_and_push(seen, ptr));
-            if is_cycle {
-                return match kind {
-                    crate::value::ArrayKind::Array | crate::value::ArrayKind::Shaped => {
-                        "[...]".to_string()
-                    }
-                    _ => "(...)".to_string(),
-                };
-            }
-            let inner = items.iter().map(gist_value).collect::<Vec<_>>().join(" ");
-            SEEN_PTRS.with(|seen| pop_ptr(seen, ptr));
-            match kind {
-                crate::value::ArrayKind::Array
-                | crate::value::ArrayKind::Shaped
-                | crate::value::ArrayKind::Lazy
-                | crate::value::ArrayKind::ItemArray => {
-                    // .gist does NOT show the `$` prefix — only .raku does.
-                    format!("[{}]", inner)
-                }
-                crate::value::ArrayKind::List | crate::value::ArrayKind::ItemList => {
-                    format!("({})", inner)
-                }
-            }
-        }
-        Value::Hash(items) => {
-            let ptr = std::sync::Arc::as_ptr(items) as usize;
-            let is_cycle = SEEN_PTRS.with(|seen| check_and_push(seen, ptr));
-            if is_cycle {
-                return "{...}".to_string();
-            }
-            let mut sorted_keys: Vec<&String> = items.keys().collect();
-            sorted_keys.sort();
-            let parts: Vec<String> = sorted_keys
-                .iter()
-                .map(|k| format!("{} => {}", k, gist_value(&items[*k])))
-                .collect();
-            SEEN_PTRS.with(|seen| pop_ptr(seen, ptr));
-            format!("{{{}}}", parts.join(", "))
-        }
-        Value::Set(..) | Value::Bag(..) | Value::Mix(..) => {
-            // Set/Bag/Mix gist shows the type-name wrapper, e.g. `Set(a b c)`;
-            // their `.Str` (the `_` fall-through) shows only the bare elements.
-            setbagmix_gist(value).unwrap_or_else(|| value.to_string_value())
-        }
-        Value::Pair(k, v) => format!("{} => {}", k, gist_value(v)),
-        Value::ValuePair(k, v) => format!("{} => {}", gist_value(k), gist_value(v)),
-        Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) | Value::Slip(items) => {
-            format!(
-                "({})",
-                items.iter().map(gist_value).collect::<Vec<_>>().join(" ")
-            )
-        }
-        Value::Version { .. } => format!("v{}", value.to_string_value()),
-        Value::Nil => "Nil".to_string(),
-        // Range.gist is identical to Range.raku in Rakudo: it shows the range
-        // notation (not the expanded elements), numeric endpoints render plainly,
-        // string endpoints are quoted (`"a".."c"`), `i64::MAX`/Whatever endpoints
-        // render as `Inf`/`-Inf`, and `0..^N` uses the `^N` short form. Delegate
-        // to the raku renderer so all of this stays in sync.
-        Value::Range(..)
-        | Value::RangeExcl(..)
-        | Value::RangeExclStart(..)
-        | Value::RangeExclBoth(..)
-        | Value::GenericRange { .. } => crate::builtins::methods_0arg::raku_repr::raku_value(value),
-        // A Match nested inside a container (e.g. the values of `$/.caps` or a
-        // `m:g//` result list) must still gist as `｢matched｣` plus its sub-
-        // captures, matching `Match.gist`. The generic Instance fall-through
-        // below would otherwise stringify it to the bare matched text.
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Match" => match_gist(&(attributes).as_map(), 0),
-        // `$(...)` itemized container: `.gist` never shows the itemization sigil,
-        // so it gists exactly like its inner value (`${a=>1}.gist` → `{a => 1}`).
-        Value::Scalar(inner) => gist_value(inner),
-        // An allomorph (IntStr/NumStr/…) gists as its preserved source string
-        // (`<1e3>.gist` → `1e3`, not the inner Num's `1000`); a general mixin
-        // gists via its inner value.
-        Value::Mixin(inner, mixins) => {
-            if crate::value::types::allomorph_type_name(inner, mixins).is_some()
-                && let Some(str_val) = mixins.get("Str")
-            {
-                str_val.to_string_value()
-            } else {
-                gist_value(inner)
-            }
-        }
-        // A WhateverCode (`*+1`, `*.abs`) gists as `WhateverCode.new`, not the
-        // empty string its bare closure stringification would yield.
-        Value::Sub(data)
-            if matches!(
-                data.env.get("__mutsu_callable_type"),
-                Some(Value::Str(kind)) if kind.as_str() == "WhateverCode"
-            ) =>
-        {
-            "WhateverCode.new".to_string()
-        }
-        _ => value.to_string_value(),
-    }
-}
-
-/// Render a Match's `.gist`: the corner-quoted matched text followed by its
-/// positional and named sub-captures, each on its own indented line, ordered by
-/// the capture's start position (`from`) and nested recursively. This mirrors
-/// Rakudo's `Match.gist`:
-///
-/// ```text
-/// ｢a1b2｣
-///  0 => ｢a1｣
-///   0 => ｢a｣
-///   1 => ｢1｣
-///  0 => ｢b2｣
-///   0 => ｢b｣
-///   1 => ｢2｣
-/// ```
-///
-/// A quantified capture (`(\w)+`) is a list of Match values, each emitted under
-/// the same index. `depth` controls indentation (one leading space per level).
-pub(crate) fn match_gist(attributes: &HashMap<String, Value>, depth: usize) -> String {
-    let matched = attributes
-        .get("str")
-        .map(|s| s.to_string_value())
-        .unwrap_or_default();
-    let mut out = format!("\u{FF62}{}\u{FF63}", matched);
-
-    // Flatten captures into (from, label, match-value) entries so a quantified
-    // capture contributes one entry per repetition, then order by match start
-    // position (named and positional interleave by position).
-    let mut entries: Vec<(i64, String, Value)> = Vec::new();
-    let push_capture = |label: &str, value: &Value, entries: &mut Vec<(i64, String, Value)>| {
-        match value {
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "Match" => {
-                entries.push((
-                    match_from(&(attributes).as_map()),
-                    label.to_string(),
-                    value.clone(),
-                ));
-            }
-            // Quantified capture: a list of Match values under one index.
-            Value::Array(items, _) => {
-                for item in items.iter() {
-                    if let Value::Instance {
-                        class_name,
-                        attributes,
-                        ..
-                    } = item
-                        && class_name == "Match"
-                    {
-                        entries.push((
-                            match_from(&(attributes).as_map()),
-                            label.to_string(),
-                            item.clone(),
-                        ));
-                    }
-                }
-            }
-            Value::Seq(items) | Value::Slip(items) => {
-                for item in items.iter() {
-                    if let Value::Instance {
-                        class_name,
-                        attributes,
-                        ..
-                    } = item
-                        && class_name == "Match"
-                    {
-                        entries.push((
-                            match_from(&(attributes).as_map()),
-                            label.to_string(),
-                            item.clone(),
-                        ));
-                    }
-                }
-            }
-            _ => {}
-        }
-    };
-
-    if let Some(Value::Array(list, _)) = attributes.get("list") {
-        for (i, cap) in list.iter().enumerate() {
-            push_capture(&i.to_string(), cap, &mut entries);
-        }
-    }
-    if let Some(Value::Hash(named)) = attributes.get("named") {
-        let mut keys: Vec<&String> = named.keys().collect();
-        keys.sort();
-        for k in keys {
-            if let Some(v) = named.get(k) {
-                push_capture(k, v, &mut entries);
-            }
-        }
-    }
-    entries.sort_by_key(|(from, _, _)| *from);
-
-    let indent = " ".repeat(depth + 1);
-    for (_, label, val) in entries {
-        if let Value::Instance { attributes, .. } = &val {
-            out.push_str(&format!(
-                "\n{}{} => {}",
-                indent,
-                label,
-                match_gist(&(attributes).as_map(), depth + 1)
-            ));
-        }
-    }
-    out
-}
-
-/// The `from` (start offset) of a Match's attributes, or 0 when absent.
-fn match_from(attributes: &HashMap<String, Value>) -> i64 {
-    match attributes.get("from") {
-        Some(Value::Int(n)) => *n,
-        _ => 0,
-    }
-}
-
-pub(crate) fn is_known_type_constraint(constraint: &str) -> bool {
-    if super::native_types::is_native_int_type(constraint) {
-        return true;
-    }
-    // Handle parameterized types like Array[Int], Hash[Str], etc.
-    if let Some(base) = constraint.split('[').next()
-        && constraint.contains('[')
-        && constraint.ends_with(']')
-        && is_known_type_constraint(base)
-    {
-        return true;
-    }
-    matches!(
-        constraint,
-        "Int"
-            | "UInt"
-            | "Num"
-            | "Str"
-            | "Bool"
-            | "Array"
-            | "Hash"
-            | "Rat"
-            | "FatRat"
-            | "Complex"
-            | "atomicint"
-            | "int"
-            | "num"
-            | "num32"
-            | "num64"
-            | "str"
-            // Core types
-            | "Mu"
-            | "Any"
-            | "Cool"
-            | "Nil"
-            | "Pair"
-            | "Map"
-            | "List"
-            | "Slip"
-            | "Seq"
-            | "Range"
-            | "Set"
-            | "SetHash"
-            | "Bag"
-            | "BagHash"
-            | "Mix"
-            | "MixHash"
-            | "QuantHash"
-            | "Blob"
-            | "Buf"
-            // Native sized buffer aliases: blob8 == Blob[uint8], buf16 == Buf[uint16], etc.
-            // (utf8/utf16/utf32 are already listed below.)
-            | "blob"
-            | "buf"
-            | "blob8"
-            | "blob16"
-            | "blob32"
-            | "blob64"
-            | "buf8"
-            | "buf16"
-            | "buf32"
-            | "buf64"
-            | "Junction"
-            | "Match"
-            | "Regex"
-            | "Block"
-            | "Code"
-            | "Sub"
-            | "Method"
-            | "Submethod"
-            | "Routine"
-            | "Capture"
-            | "Signature"
-            | "Parameter"
-            | "Whatever"
-            | "WhateverCode"
-            | "HyperWhatever"
-            | "Format"
-            | "Version"
-            | "Date"
-            | "DateTime"
-            | "Instant"
-            | "Duration"
-            | "Failure"
-            | "Exception"
-            | "Backtrace"
-            | "Scalar"
-            | "Proxy"
-            | "Label"
-            | "Grammar"
-            // Allomorphs
-            | "Allomorph"
-            | "IntStr"
-            | "NumStr"
-            | "RatStr"
-            | "ComplexStr"
-            // Meta types
-            | "Attribute"
-            | "AST"
-            | "Compiler"
-            | "Distro"
-            | "Kernel"
-            | "Macro"
-            | "ObjAt"
-            | "ValueObjAt"
-            | "Stash"
-            | "PseudoStash"
-            | "Raku"
-            | "Perl"
-            | "VM"
-            | "Variable"
-            // Array-like
-            | "array"
-            | "Hashray"
-            // Numeric subtypes
-            | "byte"
-            | "uint"
-            // IO types
-            | "IO"
-            // Ref types
-            | "IntAttrRef"
-            | "IntLexRef"
-            | "IntPosRef"
-            | "NumAttrRef"
-            | "NumLexRef"
-            | "NumPosRef"
-            | "StrAttrRef"
-            | "StrLexRef"
-            | "StrPosRef"
-            | "UIntAttrRef"
-            | "UIntLexRef"
-            | "UIntPosRef"
-            | "ScalarVAR"
-            // String types
-            | "StrDistance"
-            // Iteration
-            | "IterationBuffer"
-            // Deprecation
-            | "Deprecation"
-            | "Distribution"
-            // Operators
-            | "Operator"
-            | "OperatorProperties"
-            // Unicode types
-            | "Unicode"
-            | "Uni"
-            | "NFC"
-            | "NFD"
-            | "NFKC"
-            | "NFKD"
-            | "utf8"
-            | "utf16"
-            | "utf32"
-            | "SignedBlob"
-            | "UnsignedBlob"
-            // Concurrent types
-            | "Cancellation"
-            | "Channel"
-            | "CurrentThreadScheduler"
-            | "Promise"
-            | "Semaphore"
-            | "Supplier"
-            | "Tap"
-            | "Thread"
-            | "ThreadPoolScheduler"
-            // Hyper/Parallel
-            | "HyperConfiguration"
-            | "HyperSeq"
-            | "ParallelSequence"
-            // Slang
-            | "Slang"
-            // Lock
-            | "Lock"
-            // CallFrame
-            | "CallFrame"
-            // CompUnit
-            | "CompUnit"
-    )
-}
-
-/// Check if a compound (`::`-separated) name is a known Raku standard type.
-pub(crate) fn is_known_compound_type(name: &str) -> bool {
-    matches!(
-        name,
-        "Backtrace::Frame"
-            | "CompUnit::DependencySpecification"
-            | "CompUnit::RepositoryRegistry"
-            | "CompUnit::Repository::FileSystem"
-            | "CompUnit::Repository::Installable"
-            | "CompUnit::Repository::Installation"
-            | "CompUnit::Repository::Locally"
-            | "CompUnit::Repository::NQP"
-            | "CompUnit::Repository::Perl5"
-            | "Distribution"
-            | "Distribution::Hash"
-            | "Distribution::Installation"
-            | "Distribution::Path"
-            | "IO::ArgFiles"
-            | "IO::CatHandle"
-            | "IO::Handle"
-            | "IO::Notification"
-            | "IO::Path"
-            | "IO::Path::Cygwin"
-            | "IO::Path::QNX"
-            | "IO::Path::Unix"
-            | "IO::Path::Win32"
-            | "IO::Pipe"
-            | "IO::Socket::INET"
-            | "IO::Socket::Async"
-            | "IO::Spec"
-            | "IO::Spec::Cygwin"
-            | "IO::Spec::QNX"
-            | "IO::Spec::Unix"
-            | "IO::Spec::Win32"
-            | "IO::Special"
-            | "Metamodel::Primitives"
-            | "Pod::Block"
-            | "Pod::Block::Code"
-            | "Pod::Block::Comment"
-            | "Pod::Block::Declarator"
-            | "Pod::Block::Named"
-            | "Pod::Block::Para"
-            | "Pod::Block::Table"
-            | "Pod::Config"
-            | "Pod::Defn"
-            | "Pod::FormattingCode"
-            | "Pod::Heading"
-            | "Pod::Item"
-            | "Pod::Raw"
-            | "Proc::Async"
-            // Exception types
-            | "X::AdHoc"
-            | "X::Adverb"
-            | "X::Anon::Augment"
-            | "X::Anon::Multi"
-            | "X::Assignment::RO"
-            | "X::Assignment::RO::Comp"
-            | "X::Attribute::NoPackage"
-            | "X::Attribute::Package"
-            | "X::Attribute::Regex"
-            | "X::Attribute::Required"
-            | "X::Attribute::Undeclared"
-            | "X::Augment::NoSuchType"
-            | "X::Backslash::NonVariableDollar"
-            | "X::Backslash::UnrecognizedSequence"
-            | "X::Bind"
-            | "X::Bind::NativeType"
-            | "X::Bind::Slice"
-            | "X::Bind::ZenSlice"
-            | "X::Buf::AsStr"
-            | "X::Buf::Pack"
-            | "X::Buf::Pack::NonASCII"
-            | "X::Caller::NotDynamic"
-            | "X::Cannot::Empty"
-            | "X::Cannot::Lazy"
-            | "X::Cannot::New"
-            | "X::Channel::ReceiveOnClosed"
-            | "X::Channel::SendOnClosed"
-            | "X::Comp::AdHoc"
-            | "X::Comp::BeginTime"
-            | "X::Comp::Group"
-            | "X::Comp::NYI"
-            | "X::Comp::Trait::NotOnNative"
-            | "X::Comp::Trait::Scope"
-            | "X::Comp::Trait::Unknown"
-            | "X::Composition::NotComposable"
-            | "X::Constructor::Positional"
-            | "X::ControlFlow"
-            | "X::ControlFlow::Return"
-            | "X::DateTime::InvalidDeltaUnit"
-            | "X::DateTime::TimezoneClash"
-            | "X::Declaration::OurScopeInRole"
-            | "X::Declaration::Scope"
-            | "X::Declaration::Scope::Multi"
-            | "X::Does::TypeObject"
-            | "X::Dynamic::NotFound"
-            | "X::Dynamic::Package"
-            | "X::Dynamic::Postdeclaration"
-            | "X::Eval::NoSuchLang"
-            | "X::Export::NameClash"
-            | "X::EXPORTHOW::Conflict"
-            | "X::EXPORTHOW::InvalidDirective"
-            | "X::EXPORTHOW::NothingToSupersede"
-            | "X::Hash::Store::OddNumber"
-            | "X::HyperOp::Infinite"
-            | "X::HyperOp::NonDWIM"
-            | "X::HyperWhatever::Multiple"
-            | "X::IllegalOnFixedDimensionArray"
-            | "X::IO::Chdir"
-            | "X::IO::Chmod"
-            | "X::IO::Copy"
-            | "X::IO::Cwd"
-            | "X::IO::Dir"
-            | "X::IO::Directory"
-            | "X::IO::DoesNotExist"
-            | "X::IO::Link"
-            | "X::IO::Mkdir"
-            | "X::IO::NotAChild"
-            | "X::IO::NotAFile"
-            | "X::IO::Resolve"
-            | "X::IO::Rename"
-            | "X::IO::Rmdir"
-            | "X::IO::Symlink"
-            | "X::IO::Unlink"
-            | "X::Immutable"
-            | "X::Import::MissingSymbols"
-            | "X::Import::NoSuchTag"
-            | "X::Import::OnlystarProto"
-            | "X::Import::Positional"
-            | "X::Import::Redeclaration"
-            | "X::Inheritance::NotComposed"
-            | "X::Inheritance::SelfInherit"
-            | "X::Inheritance::UnknownParent"
-            | "X::Inheritance::Unsupported"
-            | "X::InvalidType"
-            | "X::InvalidTypeSmiley"
-            | "X::Item"
-            | "X::Localizer::NoContainer"
-            | "X::Lock::ConditionVariable::New"
-            | "X::Match::Bool"
-            | "X::Method::InvalidQualifier"
-            | "X::Method::NotFound"
-            | "X::Method::Private::Permission"
-            | "X::Method::Private::Unqualified"
-            | "X::Mixin::NotComposable"
-            | "X::Multi::Ambiguous"
-            | "X::Multi::NoMatch"
-            | "X::NQP::NotFound"
-            | "X::NYI"
-            | "X::NYI::Available"
-            | "X::NYI::BigInt"
-            | "X::NoDispatcher"
-            | "X::NoSuchSymbol"
-            | "X::NotEnoughDimensions"
-            | "X::NotParametric"
-            | "X::Numeric::DivideByZero"
-            | "X::Numeric::Real"
-            | "X::Numeric::Underflow"
-            | "X::Obsolete"
-            | "X::OutOfRange"
-            | "X::Package::Stubbed"
-            | "X::Package::UseLib"
-            | "X::Pairup::OddNumber"
-            | "X::Parameter::AfterDefault"
-            | "X::Parameter::BadType"
-            | "X::Parameter::Default"
-            | "X::Parameter::Default::TypeCheck"
-            | "X::Parameter::InvalidType"
-            | "X::Parameter::MultipleTypeConstraints"
-            | "X::Parameter::Placeholder"
-            | "X::Parameter::RW"
-            | "X::Parameter::Twigil"
-            | "X::Parameter::TypedSlurpy"
-            | "X::Parameter::WrongOrder"
-            | "X::Phaser::Multiple"
-            | "X::Phaser::PrePost"
-            | "X::PhaserExceptions"
-            | "X::Placeholder::Attribute"
-            | "X::Placeholder::Block"
-            | "X::Placeholder::Mainline"
-            | "X::Placeholder::NonPlaceholder"
-            | "X::PoisonedAlias"
-            | "X::Pragma::CannotPrecomp"
-            | "X::Pragma::CannotWhat"
-            | "X::Pragma::MustOneOf"
-            | "X::Pragma::NoArgs"
-            | "X::Pragma::OnlyOne"
-            | "X::Pragma::UnknownArg"
-            | "X::Proc::Async::AlreadyStarted"
-            | "X::Proc::Async::CharsOrBytes"
-            | "X::Proc::Async::MustBeStarted"
-            | "X::Proc::Async::OpenForWriting"
-            | "X::Proc::Async::TapBeforeSpawn"
-            | "X::Proc::Unsuccessful"
-            | "X::Promise::CauseOnlyValidOnBroken"
-            | "X::Promise::Combinator"
-            | "X::Promise::Vowed"
-            | "X::PseudoPackage::InDeclaration"
-            | "X::Range::InvalidArg"
-            | "X::Redeclaration"
-            | "X::Redeclaration::Outer"
-            | "X::Role::Initialization"
-            | "X::Role::Parametric::NoSuchCandidate"
-            | "X::Routine::Unwrap"
-            | "X::Seq::Consumed"
-            | "X::Seq::NotIndexable"
-            | "X::Sequence::Deduction"
-            | "X::Set::Coerce"
-            | "X::Signature::NameClash"
-            | "X::Signature::Placeholder"
-            | "X::Str::Match::x"
-            | "X::Str::Numeric"
-            | "X::Str::Trans::IllegalKey"
-            | "X::Str::Trans::InvalidArg"
-            | "X::StubCode"
-            | "X::Subscript::Negative"
-            | "X::Supply::Migrate::Needs"
-            | "X::Syntax::AddCategorical::TooFewParts"
-            | "X::Syntax::AddCategorical::TooManyParts"
-            | "X::Syntax::Adverb"
-            | "X::Syntax::Argument::MOPMacro"
-            | "X::Syntax::Augment::Illegal"
-            | "X::Syntax::Augment::WithoutMonkeyTyping"
-            | "X::Syntax::BlockGobbled"
-            | "X::Syntax::CannotMeta"
-            | "X::Syntax::Comment::Embedded"
-            | "X::Syntax::ConditionalOperator::PrecedenceTooLoose"
-            | "X::Syntax::ConditionalOperator::SecondPartGobbled"
-            | "X::Syntax::ConditionalOperator::SecondPartInvalid"
-            | "X::Syntax::Confused"
-            | "X::Syntax::DuplicatedPrefix"
-            | "X::Syntax::Extension::Category"
-            | "X::Syntax::Extension::Null"
-            | "X::Syntax::Extension::SpecialForm"
-            | "X::Syntax::Extension::TooComplex"
-            | "X::Syntax::InfixInTermPosition"
-            | "X::Syntax::KeywordAsFunction"
-            | "X::Syntax::Malformed"
-            | "X::Syntax::Malformed::Elsif"
-            | "X::Syntax::Missing"
-            | "X::Syntax::Name::Null"
-            | "X::Syntax::NegatedPair"
-            | "X::Syntax::NoSelf"
-            | "X::Syntax::NonAssociative"
-            | "X::Syntax::Number::IllegalDecimal"
-            | "X::Syntax::Number::RadixOutOfRange"
-            | "X::Syntax::P5"
-            | "X::Syntax::Perl5Var"
-            | "X::Syntax::Pod::BeginWithoutEnd"
-            | "X::Syntax::Pod::BeginWithoutIdentifier"
-            | "X::Syntax::Regex::Adverb"
-            | "X::Syntax::Regex::MalformedRange"
-            | "X::Syntax::Regex::NonQuantifiable"
-            | "X::Syntax::Regex::QuantifierValue"
-            | "X::Syntax::Regex::NullRegex"
-            | "X::Syntax::Regex::SolitaryBacktrackControl"
-            | "X::Syntax::Regex::SolitaryQuantifier"
-            | "X::Syntax::Regex::SpacesInBareRange"
-            | "X::Syntax::Regex::UnrecognizedMetachar"
-            | "X::Syntax::Regex::UnrecognizedModifier"
-            | "X::Syntax::Regex::Unspace"
-            | "X::Syntax::Regex::Unterminated"
-            | "X::Syntax::Reserved"
-            | "X::Syntax::Self::WithoutObject"
-            | "X::Syntax::Signature::InvocantMarker"
-            | "X::Syntax::Term::MissingInitializer"
-            | "X::Syntax::UnlessElse"
-            | "X::Syntax::Variable::BadType"
-            | "X::Syntax::Variable::ConflictingTypes"
-            | "X::Syntax::Variable::IndirectDeclaration"
-            | "X::Syntax::Variable::Initializer"
-            | "X::Syntax::Variable::Match"
-            | "X::Syntax::Variable::MissingInitializer"
-            | "X::Syntax::Variable::Numeric"
-            | "X::Syntax::Variable::Twigil"
-            | "X::Syntax::VirtualCall"
-            | "X::Temporal::InvalidFormat"
-            | "X::TooLateForREPR"
-            | "X::Trait::NotOnNative"
-            | "X::Trait::Scope"
-            | "X::Trait::Unknown"
-            | "X::TypeCheck"
-            | "X::TypeCheck::Argument"
-            | "X::TypeCheck::Assignment"
-            | "X::TypeCheck::Binding"
-            | "X::TypeCheck::Return"
-            | "X::TypeCheck::Splice"
-            | "X::Undeclared"
-            | "X::Undeclared::Symbols"
-            | "X::UnitScope::Invalid"
-            | "X::UnitScope::TooLate"
-            | "X::Value::Dynamic"
-            | "X::Worry"
-            | "X::Worry::P5"
-            | "X::Worry::P5::BackReference"
-            | "X::Worry::P5::LeadingZero"
-            | "X::Worry::P5::Reference"
-    )
-}
-
-pub(crate) fn value_type_name(value: &Value) -> &'static str {
-    match value {
-        Value::Int(_) => "Int",
-        Value::BigInt(_) => "Int",
-        Value::Num(_) => "Num",
-        Value::Str(_) => "Str",
-        Value::Bool(_) => "Bool",
-        Value::Array(_, kind) if kind.is_real_array() => "Array",
-        Value::Array(_, _) => "List",
-        Value::LazyList(_) => "Array",
-        Value::Hash(_) => "Hash",
-        Value::Range(_, _)
-        | Value::RangeExcl(_, _)
-        | Value::RangeExclStart(_, _)
-        | Value::RangeExclBoth(_, _)
-        | Value::GenericRange { .. } => "Range",
-        Value::Pair(_, _) | Value::ValuePair(_, _) => "Pair",
-        Value::Rat(_, _) => "Rat",
-        Value::FatRat(_, _) => "FatRat",
-        Value::BigRat(_, _) => "Rat",
-        Value::Complex(_, _) => "Complex",
-        Value::Set(_, is_mutable) => {
-            if *is_mutable {
-                "SetHash"
-            } else {
-                "Set"
-            }
-        }
-        Value::Bag(_, is_mutable) => {
-            if *is_mutable {
-                "BagHash"
-            } else {
-                "Bag"
-            }
-        }
-        Value::Mix(_, is_mutable) => {
-            if *is_mutable {
-                "MixHash"
-            } else {
-                "Mix"
-            }
-        }
-        Value::Nil => "Any",
-        Value::Sub(data) => match data.env.get("__mutsu_callable_type") {
-            Some(Value::Str(kind)) if kind.as_str() == "Method" => "Method",
-            Some(Value::Str(kind)) if kind.as_str() == "Submethod" => "Submethod",
-            Some(Value::Str(kind)) if kind.as_str() == "WhateverCode" => "WhateverCode",
-            Some(Value::Str(kind)) if kind.as_str() == "Block" => "Block",
-            _ => {
-                if data.is_bare_block {
-                    "Block"
-                } else {
-                    "Sub"
-                }
-            }
-        },
-        Value::WeakSub(_) => "Sub",
-        Value::Routine { is_regex, .. } => {
-            if *is_regex {
-                "Regex"
-            } else {
-                "Routine"
-            }
-        }
-        Value::Package(_) => "Package",
-        Value::CompUnitDepSpec { .. } => "Any",
-        Value::Enum { .. } => "Int",
-        Value::Instance { .. } => "Any",
-        Value::Junction { .. } => "Junction",
-        Value::Regex(_) | Value::RegexWithAdverbs { .. } => "Regex",
-        Value::Version { .. } => "Version",
-        Value::Seq(_) => "Seq",
-        Value::HyperSeq(_) => "HyperSeq",
-        Value::RaceSeq(_) => "RaceSeq",
-        Value::Slip(_) => "Slip",
-        Value::Promise(_) => "Promise",
-        Value::Channel(_) => "Channel",
-        Value::Whatever => "Whatever",
-        Value::HyperWhatever => "HyperWhatever",
-        Value::Capture { positional, named } => {
-            // Unwrap internal VarRef captures used by the VM for rw binding
-            if positional.is_empty()
-                && named
-                    .get("__mutsu_varref_name")
-                    .is_some_and(|v| matches!(v, Value::Str(_)))
-                && let Some(inner) = named.get("__mutsu_varref_value")
-            {
-                return value_type_name(inner);
-            }
-            "Capture"
-        }
-        Value::Uni(u) => match u.form.as_str() {
-            "NFC" => "NFC",
-            "NFD" => "NFD",
-            "NFKC" => "NFKC",
-            "NFKD" => "NFKD",
-            _ => "Uni",
-        },
-        Value::Mixin(inner, mixins) => {
-            if mixins.contains_key("Str") {
-                match inner.as_ref() {
-                    Value::Int(_) | Value::BigInt(_) => "IntStr",
-                    Value::Num(_) => "NumStr",
-                    Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _) => "RatStr",
-                    Value::Complex(_, _) => "ComplexStr",
-                    _ => value_type_name(inner),
-                }
-            } else {
-                value_type_name(inner)
-            }
-        }
-        Value::Proxy { .. } => "Proxy",
-        Value::ParametricRole { .. } => "Package",
-        Value::CustomType { .. } => "CustomType",
-        Value::CustomTypeInstance(_) => "CustomTypeInstance",
-        Value::Scalar(inner) => value_type_name(inner),
-        Value::LazyThunk(thunk_data) => {
-            let cache = thunk_data.cache.lock().unwrap();
-            if let Some(ref cached) = *cache {
-                // Leak the type name since we need a 'static str
-                // This is fine because type names are a small finite set
-                return value_type_name(cached);
-            }
-            "Scalar"
-        }
-        Value::LazyIoLines { .. } => "Seq",
-        Value::HashEntryRef { .. } => value_type_name(&value.hash_entry_read()),
-        Value::ContainerRef(_) => value.with_deref(value_type_name),
-    }
-}
-
-pub(crate) fn is_chain_comparison_op(op: &str) -> bool {
-    matches!(
-        op,
-        "==" | "!="
-            | "<"
-            | ">"
-            | "<="
-            | ">="
-            | "==="
-            | "!=="
-            | "=:="
-            | "eqv"
-            | "eq"
-            | "ne"
-            | "lt"
-            | "gt"
-            | "le"
-            | "ge"
-            | "before"
-            | "after"
-            | "~~"
-            | "!~~"
-            | "cmp"
-            | "leg"
-            | "<=>"
-            | "%%"
-            | "!%%"
-    ) || matches!(
-        op.strip_prefix('!'),
-        Some("==")
-            | Some("===")
-            | Some("=:=")
-            | Some("eqv")
-            | Some("eq")
-            | Some("ne")
-            | Some("lt")
-            | Some("gt")
-            | Some("le")
-            | Some("ge")
-            | Some("before")
-            | Some("after")
-            | Some("cmp")
-            | Some("leg")
-            | Some("<=>")
-    )
-}
-
-pub(crate) fn reduction_identity(op: &str) -> Value {
-    if is_chain_comparison_op(op) {
-        return Value::Bool(true);
-    }
-    match op {
-        "+" | "-" | "+|" | "+^" => Value::Int(0),
-        "*" | "**" => Value::Int(1),
-        "+&" => Value::Int(-1), // +^0 (all bits set)
-        "~" | "~|" | "~^" => Value::str(String::new()),
-        "&&" | "and" | "?&" => Value::Bool(true),
-        "||" | "or" | "?|" | "^^" => Value::Bool(false),
-        "?^" => Value::Bool(false),
-        "//" | "orelse" => Value::Package(Symbol::intern("Any")),
-        "andthen" => Value::Bool(true),
-        "xor" => Value::Bool(false),
-        "min" => Value::Num(f64::INFINITY),
-        "max" => Value::Num(f64::NEG_INFINITY),
-        // Junction operators
-        "&" => Value::Junction {
-            kind: crate::value::JunctionKind::All,
-            values: std::sync::Arc::new(Vec::new()),
-        },
-        "|" => Value::Junction {
-            kind: crate::value::JunctionKind::Any,
-            values: std::sync::Arc::new(Vec::new()),
-        },
-        "^" => Value::Junction {
-            kind: crate::value::JunctionKind::One,
-            values: std::sync::Arc::new(Vec::new()),
-        },
-        // Set operators
-        "(-)" | "∖" | "(|)" | "∪" | "(&)" | "∩" | "(^)" | "⊖" => Value::set(HashSet::new()),
-        "(.)" | "⊍" | "(+)" | "⊎" => Value::bag(HashMap::new()),
-        // Comma: empty list
-        "," => Value::Array(
-            std::sync::Arc::new(crate::value::ArrayData::new(Vec::new())),
-            ArrayKind::List,
-        ),
-        // Zip: empty Seq (Raku returns a Seq for arity-0 Z)
-        "Z" => Value::Seq(std::sync::Arc::new(Vec::new())),
-        _ => {
-            // Hyper operator forms: >>op<<, >>op>>, <<op<<, <<op>>
-            if let Some(inner) = strip_hyper_delimiters_for_identity(op) {
-                return reduction_identity(inner);
-            }
-            Value::Nil
-        }
-    }
-}
-
-/// Strip hyper operator delimiters to find the inner operator for identity lookup.
-fn strip_hyper_delimiters_for_identity(s: &str) -> Option<&str> {
-    let after_left = s
-        .strip_prefix(">>")
-        .or_else(|| s.strip_prefix("<<"))
-        .or_else(|| s.strip_prefix('\u{00BB}'))
-        .or_else(|| s.strip_prefix('\u{00AB}'))?;
-    let inner = after_left
-        .strip_suffix(">>")
-        .or_else(|| after_left.strip_suffix("<<"))
-        .or_else(|| after_left.strip_suffix('\u{00BB}'))
-        .or_else(|| after_left.strip_suffix('\u{00AB}'))?;
-    if inner.is_empty() {
-        return None;
-    }
-    Some(inner)
-}
-
-pub(crate) fn char_idx_to_byte(text: &str, idx: usize) -> usize {
-    if idx == 0 {
-        return 0;
-    }
-    for (count, (b, _)) in text.char_indices().enumerate() {
-        if count == idx {
-            return b;
-        }
-    }
-    text.len()
-}
-
-/// Coerce a list element to a form that binds *positionally* when passed as the
-/// topic/argument of a matcher or comparator block.
-///
-/// mutsu uses the `Value` variant to distinguish a call-site named argument
-/// (`Value::Pair`, excluded from positional arity everywhere in dispatch) from a
-/// positional pair *value* (`Value::ValuePair`). Iterating a Hash yields
-/// `Value::Pair` elements, so passing such an element straight to a block (e.g.
-/// `%h.first({ .value > 1 })` / `%h.sort({ ... })`) would bind it as a *named*
-/// argument, leaving the block with zero positionals ("expected N got 0"). Map
-/// `Value::Pair` to `Value::ValuePair` so the element binds as `$_`/`$^a`.
-pub(crate) fn pair_as_positional(val: &Value) -> Value {
-    match val {
-        Value::Pair(k, v) => Value::ValuePair(Box::new(Value::str(k.clone())), v.clone()),
-        other => other.clone(),
-    }
-}
-
-/// Extract the candidate closures from a `WalkList` instance's attributes,
-/// honoring its `reversed` flag. Returns `None` if the attributes are not
-/// shaped like a WalkList.
-pub(crate) fn walk_list_candidates(attributes: &crate::value::InstanceAttrs) -> Option<Vec<Value>> {
-    let map = attributes.as_map();
-    let Some(Value::Array(items, ..)) = map.get("candidates") else {
-        return None;
-    };
-    let mut cands = items.to_vec();
-    if matches!(map.get("reversed"), Some(Value::Bool(true))) {
-        cands.reverse();
-    }
-    Some(cands)
-}
-
-pub(crate) fn value_to_list(val: &Value) -> Vec<Value> {
-    match val {
-        Value::Array(items, kind) if kind.is_itemized() => vec![val.clone()],
-        Value::Array(items, ..) => items.to_vec(),
-        Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) => items.to_vec(),
-        Value::LazyList(ll) => ll.cache.lock().unwrap().clone().unwrap_or_default(),
-        // An itemized hash (`item %h` / `$(%h)`) is a single list element and does
-        // NOT flatten to its pairs (mirrors the itemized-Array arm above).
-        Value::Hash(items) if items.itemized => vec![val.clone()],
-        // `typed_pair` decontainerizes element cells so the pair value matches a
-        // `%h<k>` read / `.values` (see t/bind-hash-value-pairs.t).
-        Value::Hash(items) => items
-            .iter()
-            .map(|(k, v)| items.typed_pair(k, v.clone()))
-            .collect(),
-        Value::Range(a, b) => {
-            let end = (*b).min(*a + MAX_RANGE_EXPAND);
-            (*a..=end).map(Value::Int).collect()
-        }
-        Value::RangeExcl(a, b) => {
-            let end = (*b).min(*a + MAX_RANGE_EXPAND);
-            (*a..end).map(Value::Int).collect()
-        }
-        Value::RangeExclStart(a, b) => {
-            let start = *a + 1;
-            let end = (*b).min(start + MAX_RANGE_EXPAND);
-            (start..=end).map(Value::Int).collect()
-        }
-        Value::RangeExclBoth(a, b) => {
-            let start = *a + 1;
-            let end = (*b).min(start + MAX_RANGE_EXPAND);
-            (start..end).map(Value::Int).collect()
-        }
-        Value::GenericRange {
-            start,
-            end,
-            excl_start,
-            excl_end,
-        } => {
-            let next_numeric = |v: &Value| -> Option<Value> {
-                match v {
-                    Value::Int(i) => Some(Value::Int(i + 1)),
-                    Value::BigInt(n) => Some(Value::bigint(n.as_ref() + 1)),
-                    Value::Num(f) => Some(Value::Num(*f + 1.0)),
-                    Value::Rat(n, d) => Some(crate::value::make_rat(n + d, *d)),
-                    Value::FatRat(n, d) => Some(Value::FatRat(n + d, *d)),
-                    Value::BigRat(n, d) => {
-                        Some(Value::bigrat(n.as_ref() + d.as_ref(), (**d).clone()))
-                    }
-                    other if other.is_numeric() => Some(Value::Num(other.to_f64() + 1.0)),
-                    _ => None,
-                }
-            };
-            // String ranges: expand as character sequences
-            if let (Value::Str(a), Value::Str(b)) = (start.as_ref(), end.as_ref()) {
-                if a.chars().count() == 1 && b.chars().count() == 1 {
-                    let s = a.chars().next().unwrap() as u32;
-                    let e = b.chars().next().unwrap() as u32;
-                    let s = if *excl_start { s + 1 } else { s };
-                    if *excl_end {
-                        (s..e)
-                            .filter_map(char::from_u32)
-                            .map(|c| Value::str(c.to_string()))
-                            .collect()
-                    } else {
-                        (s..=e)
-                            .filter_map(char::from_u32)
-                            .map(|c| Value::str(c.to_string()))
-                            .collect()
-                    }
-                } else {
-                    let split_numeric_core = |s: &str| -> Option<(String, String, String)> {
-                        let mut start_byte = None;
-                        let mut end_byte = None;
-                        for (idx, ch) in s.char_indices() {
-                            if ch.is_ascii_digit() {
-                                if start_byte.is_none() {
-                                    start_byte = Some(idx);
-                                }
-                            } else if start_byte.is_some() && end_byte.is_none() {
-                                end_byte = Some(idx);
-                                break;
-                            }
-                        }
-                        let start = start_byte?;
-                        let end = end_byte.unwrap_or(s.len());
-                        if end <= start {
-                            return None;
-                        }
-                        Some((
-                            s[..start].to_string(),
-                            s[start..end].to_string(),
-                            s[end..].to_string(),
-                        ))
-                    };
-                    if let (Some((ap, an, asuf)), Some((bp, bn, bsuf))) =
-                        (split_numeric_core(a), split_numeric_core(b))
-                        && ap == bp
-                        && asuf == bsuf
-                        && let (Ok(mut n), Ok(e)) = (an.parse::<i128>(), bn.parse::<i128>())
-                    {
-                        if *excl_start {
-                            n += 1;
-                        }
-                        if n > e {
-                            return Vec::new();
-                        }
-                        let width = an.len().max(bn.len());
-                        let pad = an.starts_with('0') || bn.starts_with('0');
-                        let mut result = Vec::new();
-                        let limit = MAX_RANGE_EXPAND as usize;
-                        while n <= e && result.len() < limit {
-                            if *excl_end && n == e {
-                                break;
-                            }
-                            let digits = if pad {
-                                format!("{n:0width$}")
-                            } else {
-                                n.to_string()
-                            };
-                            result.push(Value::str(format!("{ap}{digits}{asuf}")));
-                            n += 1;
-                        }
-                        return result;
-                    }
-                    // Multi-char string ranges: use string succession
-                    let mut result = Vec::new();
-                    let mut current = if *excl_start {
-                        crate::runtime::Interpreter::string_succ(a)
-                    } else {
-                        a.to_string()
-                    };
-                    let limit = MAX_RANGE_EXPAND as usize;
-                    while current.as_str() <= b.as_str() && result.len() < limit {
-                        if *excl_end && current.as_str() == b.as_str() {
-                            break;
-                        }
-                        result.push(Value::str(current.clone()));
-                        current = crate::runtime::Interpreter::string_succ(&current);
-                    }
-                    result
-                }
-            } else if let (Value::Str(a), Value::HyperWhatever | Value::Whatever) =
-                (start.as_ref(), end.as_ref())
-            {
-                let mut result = Vec::new();
-                let mut current = if *excl_start {
-                    crate::runtime::Interpreter::string_succ(a)
-                } else {
-                    a.to_string()
-                };
-                let limit = MAX_RANGE_EXPAND as usize;
-                while result.len() < limit {
-                    result.push(Value::str(current.clone()));
-                    current = crate::runtime::Interpreter::string_succ(&current);
-                }
-                result
-            } else if let Value::Str(a) = start.as_ref() {
-                // Start is a Str — iterate as strings (preserving type).
-                // In Raku, "1"..9 produces ("1", "2", ..., "9").
-                // When end is numeric, compare numerically to determine bounds.
-                let end_is_numeric = end.as_ref().is_numeric()
-                    || matches!(end.as_ref(), Value::Whatever | Value::HyperWhatever);
-                let end_str = match end.as_ref() {
-                    Value::Str(s) => (**s).clone(),
-                    other => other.to_string_value(),
-                };
-                let end_f64 = end.as_ref().to_f64();
-                let mut result = Vec::new();
-                let mut current = if *excl_start {
-                    crate::runtime::Interpreter::string_succ(a)
-                } else {
-                    a.to_string()
-                };
-                let limit = MAX_RANGE_EXPAND as usize;
-                while result.len() < limit {
-                    let in_range = if end_is_numeric {
-                        // Compare current string numerically against end
-                        let cur_numeric = coerce_to_numeric(Value::str(current.clone()));
-                        let cur_f64 = cur_numeric.to_f64();
-                        if *excl_end {
-                            cur_f64 < end_f64
-                        } else {
-                            cur_f64 <= end_f64
-                        }
-                    } else {
-                        // String comparison
-                        if *excl_end {
-                            current.as_str() < end_str.as_str()
-                        } else {
-                            current.as_str() <= end_str.as_str()
-                        }
-                    };
-                    if !in_range {
-                        break;
-                    }
-                    result.push(Value::str(current.clone()));
-                    current = crate::runtime::Interpreter::string_succ(&current);
-                }
-                result
-            } else {
-                // Numeric GenericRange: expand using .succ semantics to preserve endpoint type.
-                let start_num = if start.is_numeric() {
-                    Some(start.as_ref().clone())
-                } else {
-                    None
-                };
-                let end_num = if matches!(end.as_ref(), Value::Whatever | Value::HyperWhatever) {
-                    Some(Value::Num(f64::INFINITY))
-                } else if end.is_numeric() {
-                    Some(end.as_ref().clone())
-                } else if let Value::Str(s) = end.as_ref() {
-                    let coerced = coerce_to_numeric(Value::str((**s).clone()));
-                    if coerced.is_numeric() {
-                        Some(coerced)
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                };
-                let (start_num, end_num) = match (start_num, end_num) {
-                    (Some(s), Some(e)) => (s, e),
-                    _ => {
-                        // Check for Date-like instances with .succ support
-                        let is_date_like = |v: &Value| -> bool {
-                            if let Value::Instance { attributes, .. } = v {
-                                attributes.contains_key("year")
-                                    && attributes.contains_key("month")
-                                    && attributes.contains_key("day")
-                                    && !attributes.contains_key("hour")
-                            } else {
-                                false
-                            }
-                        };
-                        if is_date_like(start.as_ref()) && is_date_like(end.as_ref()) {
-                            use crate::builtins::methods_0arg::temporal::{
-                                civil_to_epoch_days, date_attrs, epoch_days_to_civil,
-                                make_date_with_formatter,
-                            };
-                            let (sy, sm, sd) =
-                                if let Value::Instance { attributes, .. } = start.as_ref() {
-                                    date_attrs(&(attributes).as_map())
-                                } else {
-                                    unreachable!()
-                                };
-                            let (ey, em, ed) =
-                                if let Value::Instance { attributes, .. } = end.as_ref() {
-                                    date_attrs(&(attributes).as_map())
-                                } else {
-                                    unreachable!()
-                                };
-                            let start_days = civil_to_epoch_days(sy, sm, sd);
-                            let end_days = civil_to_epoch_days(ey, em, ed);
-                            let formatter =
-                                if let Value::Instance { attributes, .. } = start.as_ref() {
-                                    attributes.as_map().get("formatter").cloned()
-                                } else {
-                                    None
-                                };
-                            let mut result = Vec::new();
-                            let first_day = if *excl_start {
-                                start_days + 1
-                            } else {
-                                start_days
-                            };
-                            let limit = MAX_RANGE_EXPAND as usize;
-                            let mut d = first_day;
-                            while result.len() < limit {
-                                if d > end_days || (*excl_end && d == end_days) {
-                                    break;
-                                }
-                                let (y, m, dd) = epoch_days_to_civil(d);
-                                result.push(make_date_with_formatter(y, m, dd, formatter.clone()));
-                                d += 1;
-                            }
-                            return result;
-                        }
-                        return vec![val.clone()];
-                    }
-                };
-                if start_num.to_f64().is_infinite()
-                    || start_num.to_f64().is_nan()
-                    || end_num.to_f64().is_nan()
-                {
-                    vec![val.clone()]
-                } else {
-                    let mut result = Vec::new();
-                    let mut current = if *excl_start {
-                        next_numeric(&start_num).unwrap_or(start_num)
-                    } else {
-                        start_num
-                    };
-                    let limit = MAX_RANGE_EXPAND as usize;
-                    while result.len() < limit {
-                        let cmp = compare_values(&current, &end_num);
-                        if cmp > 0 || (*excl_end && cmp == 0) {
-                            break;
-                        }
-                        result.push(current.clone());
-                        let Some(next) = next_numeric(&current) else {
-                            break;
-                        };
-                        if values_identical(&next, &current) {
-                            break;
-                        }
-                        current = next;
-                    }
-                    result
-                }
-            }
-        }
-        Value::Set(items, _) => items
-            .iter()
-            .map(|s| Value::Pair(s.clone(), Box::new(Value::Bool(true))))
-            .collect(),
-        Value::Bag(items, _) => items
-            .iter()
-            .map(|(k, v)| Value::Pair(k.clone(), Box::new(Value::from_bigint(v.clone()))))
-            .collect(),
-        Value::Mix(items, _) => items
-            .iter()
-            .map(|(k, v)| Value::Pair(k.clone(), Box::new(crate::value::mix_weight_to_value(*v))))
-            .collect(),
-        Value::Slip(items) => items.to_vec(),
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } => {
-            // A WalkList flattens to its candidate closures in list context, so
-            // `my @cands = $x.WALK(...)` yields the per-level candidates.
-            if class_name.resolve() == "WalkList"
-                && let Some(items) = walk_list_candidates(attributes)
-            {
-                return items;
-            }
-            if let Some(Value::Array(items, ..)) = attributes.as_map().get("__array_items") {
-                return items.to_vec();
-            }
-            vec![val.clone()]
-        }
-        // Nil is a single scalar item in list context (e.g. `for Nil { }` does
-        // one iteration); it is not an empty list. Fall through to the scalar arm.
-        other => vec![other.clone()],
-    }
-}
+pub(crate) use coerce_containers::*;
+pub(crate) use compare::*;
+pub(crate) use errors::*;
+pub(crate) use gist::*;
+pub(crate) use list::*;
+pub(crate) use radix_numeric::*;
+pub(crate) use rat::*;
+pub(crate) use set_coerce::*;
+pub(crate) use set_ops::*;
+pub(crate) use shaped::*;
+pub(crate) use type_constraints::*;
+pub(crate) use type_misc::*;
 
 pub(crate) use super::sprintf::format_sprintf;
 pub(crate) use super::sprintf::format_sprintf_args;
 pub(crate) use super::sprintf::format_zprintf;
-
-fn parse_unicode_decimal_digits(input: &str) -> Option<(&str, String)> {
-    let mut end = 0;
-    let mut clean = String::new();
-    let mut saw_digit = false;
-    for c in input.chars() {
-        if c == '_' {
-            end += c.len_utf8();
-            continue;
-        }
-        let Some(dv) = crate::builtins::unicode::unicode_decimal_digit_value(c) else {
-            break;
-        };
-        saw_digit = true;
-        clean.push(char::from_digit(dv, 10).unwrap());
-        end += c.len_utf8();
-    }
-    if saw_digit {
-        Some((&input[end..], clean))
-    } else {
-        None
-    }
-}
-
-pub(crate) fn parse_radix_number_body(body: &str, base: u32) -> Option<Value> {
-    if !(2..=36).contains(&base) {
-        return None;
-    }
-    let compact: String = body.chars().filter(|c| !c.is_whitespace()).collect();
-    if compact.is_empty() {
-        return None;
-    }
-
-    let (digits_body, exponent_scale) =
-        if let Some((digits, exp_part)) = compact.split_once("*10**") {
-            let exp_part = exp_part.trim();
-            if exp_part.is_empty() {
-                return None;
-            }
-            let (sign, after_sign) = if let Some(rest) = exp_part.strip_prefix('+') {
-                (1_i64, rest)
-            } else if let Some(rest) = exp_part.strip_prefix('-') {
-                (-1_i64, rest)
-            } else {
-                (1_i64, exp_part)
-            };
-            let (exp_rest, exp_clean) = parse_unicode_decimal_digits(after_sign)?;
-            if !exp_rest.is_empty() {
-                return None;
-            }
-            let exp_abs: i64 = exp_clean.parse().ok()?;
-            (digits.trim(), sign * exp_abs)
-        } else {
-            (compact.trim(), 0_i64)
-        };
-    if digits_body.is_empty() {
-        return None;
-    }
-
-    let mut int_clean = String::new();
-    let mut frac_clean = String::new();
-    let mut saw_dot = false;
-    let mut saw_digit = false;
-    for c in digits_body.chars() {
-        if c == '_' {
-            continue;
-        }
-        if c == '.' {
-            if saw_dot {
-                return None;
-            }
-            saw_dot = true;
-            continue;
-        }
-        let value =
-            crate::builtins::unicode::unicode_decimal_digit_value(c).or_else(|| match c {
-                'a'..='z' => Some(10 + (c as u32 - 'a' as u32)),
-                'A'..='Z' => Some(10 + (c as u32 - 'A' as u32)),
-                'ａ'..='ｚ' => Some(10 + (c as u32 - 'ａ' as u32)),
-                'Ａ'..='Ｚ' => Some(10 + (c as u32 - 'Ａ' as u32)),
-                _ => None,
-            })?;
-        if value >= base {
-            return None;
-        }
-        let digit = char::from_digit(value, 36)?;
-        saw_digit = true;
-        if saw_dot {
-            frac_clean.push(digit);
-        } else {
-            int_clean.push(digit);
-        }
-    }
-    if !saw_digit {
-        return None;
-    }
-
-    if !saw_dot && exponent_scale == 0 {
-        if let Ok(n) = i64::from_str_radix(&int_clean, base) {
-            return Some(Value::Int(n));
-        }
-        if let Some(n) = num_bigint::BigInt::parse_bytes(int_clean.as_bytes(), base) {
-            return Some(Value::from_bigint(n));
-        }
-        return None;
-    }
-
-    let base_big = num_bigint::BigInt::from(base);
-    let int_value = if int_clean.is_empty() {
-        num_bigint::BigInt::from(0_i64)
-    } else {
-        num_bigint::BigInt::parse_bytes(int_clean.as_bytes(), base)?
-    };
-    let frac_value = if frac_clean.is_empty() {
-        num_bigint::BigInt::from(0_i64)
-    } else {
-        num_bigint::BigInt::parse_bytes(frac_clean.as_bytes(), base)?
-    };
-    let frac_scale = base_big.pow(frac_clean.len() as u32);
-    let mut numerator = int_value * &frac_scale + frac_value;
-    let mut denominator = frac_scale;
-    if exponent_scale != 0 {
-        let exp_abs = exponent_scale.unsigned_abs() as u32;
-        let scale10 = num_bigint::BigInt::from(10_u32).pow(exp_abs);
-        if exponent_scale > 0 {
-            numerator *= scale10;
-        } else {
-            denominator *= scale10;
-        }
-    }
-    Some(crate::value::make_big_rat(numerator, denominator))
-}
-
-pub(crate) fn coerce_to_numeric(val: Value) -> Value {
-    let val = val.into_descalarized();
-    match val {
-        Value::Mixin(inner, _) => coerce_to_numeric(inner.as_ref().clone()),
-        Value::Int(_)
-        | Value::BigInt(_)
-        | Value::Num(_)
-        | Value::Rat(_, _)
-        | Value::FatRat(_, _)
-        | Value::BigRat(_, _)
-        | Value::Complex(_, _) => val,
-        Value::Bool(b) => Value::Int(if b { 1 } else { 0 }),
-        Value::Enum { value, .. } => Value::Int(value.as_i64()),
-        Value::Str(ref s) => {
-            if let Some(v) = crate::runtime::str_numeric::parse_raku_str_to_numeric(s) {
-                v
-            } else {
-                Value::Int(0)
-            }
-        }
-        _ if val.as_list_items().is_some() => Value::Int(val.as_list_items().unwrap().len() as i64),
-        Value::Hash(items) => Value::Int(items.len() as i64),
-        Value::Set(items, _) => Value::Int(items.len() as i64),
-        Value::Bag(items, _) => Value::from_bigint(items.values().sum::<BigInt>()),
-        Value::Mix(items, _) => {
-            // Sort values before summing for deterministic results
-            // regardless of HashMap iteration order.
-            let mut vals: Vec<f64> = items.values().copied().collect();
-            vals.sort_by(|a, b| a.total_cmp(b));
-            let total: f64 = vals.iter().copied().fold(0.0, std::ops::Add::add);
-            if total == 0.0 && items.is_empty() {
-                Value::Int(0)
-            } else if (total - (total as i64 as f64)).abs() < f64::EPSILON {
-                Value::Int(total as i64)
-            } else {
-                Value::Num(total)
-            }
-        }
-        Value::LazyList(ll) => {
-            if let Some(count) = &ll.elems_count {
-                count.clone()
-            } else if let Some(cached) = ll.cache.lock().unwrap().as_ref() {
-                Value::Int(cached.len() as i64)
-            } else {
-                Value::Int(0)
-            }
-        }
-        Value::Range(..)
-        | Value::RangeExcl(..)
-        | Value::RangeExclStart(..)
-        | Value::RangeExclBoth(..)
-        | Value::GenericRange { .. } => Value::Int(value_to_list(&val).len() as i64),
-        Value::Nil => Value::Int(0),
-        Value::Instance {
-            ref class_name,
-            ref attributes,
-            ..
-        } if class_name == "Instant" => attributes
-            .as_map()
-            .get("value")
-            .cloned()
-            .unwrap_or(Value::Num(0.0)),
-        Value::Instance {
-            ref class_name,
-            ref attributes,
-            ..
-        } if class_name == "Duration" => attributes
-            .as_map()
-            .get("value")
-            .cloned()
-            .unwrap_or(Value::Num(0.0)),
-        Value::Instance {
-            ref class_name,
-            ref attributes,
-            ..
-        } if class_name == "Date" => {
-            let (y, m, d) =
-                crate::builtins::methods_0arg::temporal::date_attrs(&(attributes).as_map());
-            let epoch = crate::builtins::methods_0arg::temporal::civil_to_epoch_days(y, m, d);
-            Value::Int(epoch * 86400)
-        }
-        Value::Instance {
-            ref class_name,
-            ref attributes,
-            ..
-        } if class_name == "DateTime" => {
-            let (y, mo, d, h, mi, s, tz) =
-                crate::builtins::methods_0arg::temporal::datetime_attrs(&(attributes).as_map());
-            Value::Num(crate::builtins::methods_0arg::temporal::datetime_to_posix(
-                y, mo, d, h, mi, s, tz,
-            ))
-        }
-        Value::Uni(u) => Value::Int(u.text.chars().count() as i64),
-        Value::Capture { ref positional, .. } => Value::Int(positional.len() as i64),
-        _ => Value::Int(0),
-    }
-}
-
-pub(crate) fn coerce_to_set(val: &Value) -> HashSet<String> {
-    fn insert_set_elem(elems: &mut HashSet<String>, value: &Value) {
-        let pair_selected = |weight: &Value| weight.truthy() || matches!(weight, Value::Nil);
-        match value {
-            Value::Set(items, _) => {
-                elems.extend(items.iter().cloned());
-            }
-            Value::Bag(items, _) => {
-                for (k, v) in items.iter() {
-                    if v.is_positive() {
-                        elems.insert(k.clone());
-                    }
-                }
-            }
-            Value::Mix(items, _) => {
-                for (k, v) in items.iter() {
-                    if *v != 0.0 {
-                        elems.insert(k.clone());
-                    }
-                }
-            }
-            Value::Hash(items) => {
-                for (k, v) in items.iter() {
-                    if v.truthy() || matches!(v, Value::Nil) {
-                        elems.insert(k.clone());
-                    }
-                }
-            }
-            _ if value.as_list_items().is_some() => {
-                for item in value.as_list_items().unwrap().iter() {
-                    insert_set_elem(elems, item);
-                }
-            }
-            range if range.is_range() => {
-                for item in value_to_list(range) {
-                    insert_set_elem(elems, &item);
-                }
-            }
-            Value::Pair(key, weight) => {
-                if pair_selected(weight) {
-                    elems.insert(key.clone());
-                }
-            }
-            Value::ValuePair(key, weight) => {
-                if pair_selected(weight) {
-                    elems.insert(key.to_string_value());
-                }
-            }
-            other => {
-                let sv = other.to_string_value();
-                if !sv.is_empty() {
-                    elems.insert(sv);
-                }
-            }
-        }
-    }
-
-    let val = val.descalarize();
-    match val {
-        Value::Set(s, _) => s.elements.clone(),
-        Value::Bag(b, _) => {
-            let resolved = resolve_bag_tab_keys(b);
-            resolved.keys().cloned().collect()
-        }
-        Value::Mix(m, _) => m.keys().cloned().collect(),
-        Value::Hash(items) => items
-            .iter()
-            .filter_map(|(k, v)| {
-                if v.truthy() || matches!(v, Value::Nil) {
-                    Some(k.clone())
-                } else {
-                    None
-                }
-            })
-            .collect(),
-        _ if val.as_list_items().is_some() => {
-            let mut elems = HashSet::new();
-            for item in val.as_list_items().unwrap().iter() {
-                insert_set_elem(&mut elems, item);
-            }
-            elems
-        }
-        Value::Pair(_, _) | Value::ValuePair(_, _) => {
-            let mut elems = HashSet::new();
-            insert_set_elem(&mut elems, val);
-            elems
-        }
-        range if range.is_range() => {
-            let mut elems = HashSet::new();
-            for item in value_to_list(range) {
-                insert_set_elem(&mut elems, &item);
-            }
-            elems
-        }
-        _ => {
-            let mut s = HashSet::new();
-            let sv = val.to_string_value();
-            if !sv.is_empty() {
-                s.insert(sv);
-            }
-            s
-        }
-    }
-}
-
-/// Coerce a value to a QuantHash (Set/Bag/Mix) for use as a single operand to set operators.
-/// - Set/Bag/Mix pass through as-is
-/// - Hash: include keys with truthy values as Set elements
-/// - List/Array: convert items to Set (excluding Pairs with falsy value)
-/// - Pair with falsy value → empty Set
-/// - Other scalars → Set with one element
-pub(crate) fn coerce_value_to_quanthash(val: &Value) -> Value {
-    let val = val.descalarize();
-    match val {
-        Value::Set(_, _) | Value::Bag(_, _) | Value::Mix(_, _) => val.clone(),
-        Value::Hash(h) => {
-            let mut set = HashSet::new();
-            for (k, v) in h.iter() {
-                if v.truthy() {
-                    set.insert(k.clone());
-                }
-            }
-            Value::set(set)
-        }
-        _ if val.as_list_items().is_some() => {
-            let mut set = HashSet::new();
-            for item in val.as_list_items().unwrap().iter() {
-                match item {
-                    Value::Pair(k, v) => {
-                        if v.truthy() {
-                            set.insert(k.clone());
-                        }
-                    }
-                    Value::Hash(h) => {
-                        for (k, v) in h.iter() {
-                            if v.truthy() {
-                                set.insert(k.clone());
-                            }
-                        }
-                    }
-                    other => {
-                        set.insert(other.to_string_value());
-                    }
-                }
-            }
-            Value::set(set)
-        }
-        Value::Pair(k, v) => {
-            let mut set = HashSet::new();
-            if v.truthy() {
-                set.insert(k.clone());
-            }
-            Value::set(set)
-        }
-        _ => {
-            let mut set = HashSet::new();
-            let sv = val.to_string_value();
-            if !sv.is_empty() {
-                set.insert(sv);
-            }
-            Value::set(set)
-        }
-    }
-}
-
-/// Determine the promotion level for set operations: 0=Set, 1=Bag, 2=Mix
-fn set_type_level(v: &Value) -> u8 {
-    match v {
-        Value::Mix(_, _) => 2,
-        Value::Bag(_, _) => 1,
-        _ => 0,
-    }
-}
-
-/// Convert a value to a Mix-level HashMap (key → f64 count)
-fn to_mix_map(v: &Value) -> HashMap<String, f64> {
-    match v {
-        Value::Mix(m, _) => m.weights.clone(),
-        Value::Bag(b, _) => {
-            let resolved = resolve_bag_tab_keys(b);
-            resolved.into_iter().map(|(k, v)| (k, v as f64)).collect()
-        }
-        Value::Set(s, _) => s.iter().map(|k| (k.clone(), 1.0)).collect(),
-        Value::Hash(h) => {
-            let mut result = HashMap::new();
-            for (k, v) in h.iter() {
-                let w = match v {
-                    Value::Int(i) => *i as f64,
-                    Value::Num(n) => *n,
-                    Value::Rat(n, d) if *d != 0 => *n as f64 / *d as f64,
-                    Value::Bool(b) => {
-                        if *b {
-                            1.0
-                        } else {
-                            0.0
-                        }
-                    }
-                    _ => {
-                        if v.truthy() {
-                            1.0
-                        } else {
-                            0.0
-                        }
-                    }
-                };
-                if w != 0.0 {
-                    result.insert(k.clone(), w);
-                }
-            }
-            result
-        }
-        _ => {
-            // Count occurrences for list-like values (e.g. (a, a, b) → {a: 2.0, b: 1.0})
-            let items = value_to_list(v);
-            let mut result = HashMap::new();
-            for item in &items {
-                *result.entry(item.to_string_value()).or_insert(0.0f64) += 1.0;
-            }
-            result
-        }
-    }
-}
-
-/// Resolve Bag entries that use the internal "key\tweight" tab format
-/// into plain key→weight entries.
-pub(crate) fn resolve_bag_tab_keys(bag: &HashMap<String, BigInt>) -> HashMap<String, i64> {
-    let mut result = HashMap::new();
-    for (k, c) in bag.iter() {
-        let c = bigint_to_i64_sat(c);
-        if let Some((base, raw_weight)) = k.split_once('\t') {
-            let weight = match raw_weight {
-                "True" => 1i64,
-                "False" => 0,
-                _ => raw_weight.parse::<i64>().unwrap_or(1),
-            };
-            *result.entry(base.to_string()).or_insert(0) += weight * c;
-        } else {
-            *result.entry(k.clone()).or_insert(0) += c;
-        }
-    }
-    // Remove zero/negative entries for Bag semantics
-    result.retain(|_, v| *v > 0);
-    result
-}
-
-/// Convert a value to a Bag-level HashMap (key → i64 count)
-fn to_bag_map(v: &Value) -> HashMap<String, i64> {
-    match v {
-        Value::Bag(b, _) => resolve_bag_tab_keys(b),
-        Value::Set(s, _) => s.iter().map(|k| (k.clone(), 1i64)).collect(),
-        Value::Hash(h) => {
-            let mut result = HashMap::new();
-            for (k, v) in h.iter() {
-                let count = match v {
-                    Value::Int(i) => *i,
-                    Value::Num(n) => *n as i64,
-                    Value::Bool(b) => i64::from(*b),
-                    _ => i64::from(v.truthy()),
-                };
-                if count > 0 {
-                    result.insert(k.clone(), count);
-                }
-            }
-            result
-        }
-        _ => {
-            // Count occurrences for list-like values (e.g. (a, a, b) → {a: 2, b: 1})
-            let items = value_to_list(v);
-            let mut result = HashMap::new();
-            for item in &items {
-                *result.entry(item.to_string_value()).or_insert(0i64) += 1;
-            }
-            result
-        }
-    }
-}
-
-/// Whether a set operator's left operand makes the result a mutable QuantHash.
-/// Raku's set operators (`(|)`/`(&)`/`(-)`/`(^)`/`(.)`/`(+)`) take their
-/// result's mutability from the FIRST operand only: an immutable operand
-/// (Set/Bag/Mix, a list, or a type object used as an element) yields an
-/// immutable result even when a later operand is a SetHash/BagHash/MixHash.
-pub(crate) fn set_result_mutability(v: &Value) -> bool {
-    matches!(
-        v,
-        Value::Set(_, true) | Value::Bag(_, true) | Value::Mix(_, true)
-    )
-}
-
-/// Whether a value is an actual QuantHash instance (Set/Bag/Mix, mutable or
-/// not) as opposed to a list, hash, or bare type object. Used by symmetric
-/// difference, whose result stays mutable only when BOTH operands are
-/// QuantHashes (`SetHash (^) Set` -> SetHash, but `SetHash (^) <a b>` -> Set).
-pub(crate) fn is_quanthash_instance(v: &Value) -> bool {
-    matches!(v, Value::Set(_, _) | Value::Bag(_, _) | Value::Mix(_, _))
-}
-
-/// Result mutability for symmetric difference (`(^)`). Like the other set
-/// operators it follows the first operand, but at the Set level it additionally
-/// demotes to an immutable Set when the right operand is not a QuantHash
-/// (`SetHash (^) <a b>` -> Set). The demotion does NOT apply once the result is
-/// promoted to Bag/Mix level (`BagHash (^) <a b>` -> BagHash).
-pub(crate) fn set_sym_diff_mutability(left: &Value, right: &Value) -> bool {
-    set_result_mutability(left)
-        && (is_quanthash_instance(right) || matches!(left, Value::Bag(_, _) | Value::Mix(_, _)))
-}
-
-/// Overlay the given mutability onto a freshly-built set-operator result.
-pub(crate) fn with_set_mutability(result: Value, mutable: bool) -> Value {
-    match result {
-        Value::Set(d, _) => Value::Set(d, mutable),
-        Value::Bag(d, _) => Value::Bag(d, mutable),
-        Value::Mix(d, _) => Value::Mix(d, mutable),
-        other => other,
-    }
-}
-
-/// Standalone set difference: left (-) right
-/// Implements Raku type promotion: Mix > Bag > Set
-///
-/// In Raku, when the RHS is a Hash, it is treated as a Set (truthy keys
-/// with weight 1) rather than using its numeric values as weights.
-pub(crate) fn set_diff_values(left: &Value, right: &Value) -> Value {
-    // When the RHS is a Hash, coerce it to a Set for subtraction.
-    // The Hash's truthy keys each count as weight 1.
-    let right_as_set;
-    let effective_right = if matches!(right, Value::Hash(_)) {
-        right_as_set = Value::set(coerce_to_set(right));
-        &right_as_set
-    } else {
-        right
-    };
-
-    let level = set_type_level(left).max(set_type_level(effective_right));
-    match level {
-        2 => {
-            // Mix-level difference: include all keys, keep non-zero results
-            let a = to_mix_map(left);
-            let b = to_mix_map(effective_right);
-            let mut result = HashMap::new();
-            for (k, v) in &a {
-                let bv = b.get(k).copied().unwrap_or(0.0);
-                let diff = v - bv;
-                if diff != 0.0 {
-                    result.insert(k.clone(), diff);
-                }
-            }
-            for (k, v) in &b {
-                if !a.contains_key(k) && *v != 0.0 {
-                    result.insert(k.clone(), -*v);
-                }
-            }
-            Value::mix(result)
-        }
-        1 => {
-            // Bag-level difference: only positive results
-            let a = to_bag_map(left);
-            let b = to_bag_map(effective_right);
-            let mut result = HashMap::new();
-            for (k, v) in &a {
-                let bv = b.get(k).copied().unwrap_or(0);
-                if *v > bv {
-                    result.insert(k.clone(), *v - bv);
-                }
-            }
-            Value::bag(result)
-        }
-        _ => {
-            // Set-level difference
-            let a = coerce_to_set(left);
-            let b = coerce_to_set(effective_right);
-            Value::set(a.difference(&b).cloned().collect())
-        }
-    }
-}
-
-/// Standalone set intersection: left (&) right
-pub(crate) fn set_intersect_values(left: &Value, right: &Value) -> Value {
-    // Determine result type level: 0=Set, 1=Bag, 2=Mix
-    let type_level = |v: &Value| -> u8 {
-        match v {
-            Value::Mix(_, _) => 2,
-            Value::Bag(_, _) => 1,
-            _ => 0,
-        }
-    };
-    let result_level = type_level(left).max(type_level(right));
-    match result_level {
-        2 => {
-            let a = coerce_to_mix(left);
-            let b = coerce_to_mix(right);
-            let mut result = HashMap::new();
-            for (k, v) in a.iter() {
-                if let Some(bv) = b.get(k) {
-                    result.insert(k.clone(), v.min(*bv));
-                }
-            }
-            Value::mix(result)
-        }
-        1 => {
-            let a = coerce_to_bag(left);
-            let b = coerce_to_bag(right);
-            let mut result = HashMap::new();
-            for (k, v) in a.iter() {
-                if let Some(bv) = b.get(k) {
-                    result.insert(k.clone(), (*v).min(*bv));
-                }
-            }
-            Value::bag(result)
-        }
-        _ => {
-            let a = coerce_to_set(left);
-            let b = coerce_to_set(right);
-            Value::set(a.intersection(&b).cloned().collect())
-        }
-    }
-}
-
-/// Coerce a value to a Bag (HashMap<String, i64>)
-fn coerce_to_bag(val: &Value) -> HashMap<String, i64> {
-    match val {
-        Value::Bag(b, _) => resolve_bag_tab_keys(b),
-        Value::Set(s, _) => s.iter().map(|k| (k.clone(), 1)).collect(),
-        Value::Mix(m, _) => m.iter().map(|(k, v)| (k.clone(), *v as i64)).collect(),
-        _ => {
-            // Count occurrences for list-like values
-            let items = value_to_list(val);
-            let mut result = HashMap::new();
-            for item in &items {
-                *result.entry(item.to_string_value()).or_insert(0i64) += 1;
-            }
-            result
-        }
-    }
-}
-
-/// Coerce a value to a Mix (HashMap<String, f64>)
-fn coerce_to_mix(val: &Value) -> HashMap<String, f64> {
-    match val {
-        Value::Mix(m, _) => m.weights.clone(),
-        Value::Bag(b, _) => {
-            let resolved = resolve_bag_tab_keys(b);
-            resolved.into_iter().map(|(k, v)| (k, v as f64)).collect()
-        }
-        Value::Set(s, _) => s.iter().map(|k| (k.clone(), 1.0)).collect(),
-        _ => {
-            // Count occurrences for list-like values
-            let items = value_to_list(val);
-            let mut result = HashMap::new();
-            for item in &items {
-                *result.entry(item.to_string_value()).or_insert(0.0f64) += 1.0;
-            }
-            result
-        }
-    }
-}
-
-/// Standalone set symmetric difference: left (^) right
-/// Implements Raku type promotion: Mix > Bag > Set
-///
-/// For Bags: the result multiplicity for each key is |a_count - b_count|
-/// (entries with zero are dropped).
-/// For Mixes: the result weight for each key is |a_weight - b_weight|
-/// (entries with zero are dropped).
-pub(crate) fn set_sym_diff_values(left: &Value, right: &Value) -> Value {
-    let level = set_type_level(left).max(set_type_level(right));
-    match level {
-        2 => {
-            // Mix-level symmetric difference: |a - b| for each key
-            let a = to_mix_map(left);
-            let b = to_mix_map(right);
-            let mut result = HashMap::new();
-            let mut all_keys: HashSet<String> = a.keys().cloned().collect();
-            all_keys.extend(b.keys().cloned());
-            for k in all_keys {
-                let av = a.get(&k).copied().unwrap_or(0.0);
-                let bv = b.get(&k).copied().unwrap_or(0.0);
-                let diff = (av - bv).abs();
-                if diff != 0.0 {
-                    result.insert(k, diff);
-                }
-            }
-            Value::mix(result)
-        }
-        1 => {
-            // Bag-level symmetric difference: |a - b| for each key
-            let a = to_bag_map(left);
-            let b = to_bag_map(right);
-            let mut result = HashMap::new();
-            let mut all_keys: HashSet<String> = a.keys().cloned().collect();
-            all_keys.extend(b.keys().cloned());
-            for k in all_keys {
-                let av = a.get(&k).copied().unwrap_or(0);
-                let bv = b.get(&k).copied().unwrap_or(0);
-                let diff = (av - bv).unsigned_abs() as i64;
-                if diff > 0 {
-                    result.insert(k, diff);
-                }
-            }
-            Value::bag(result)
-        }
-        _ => {
-            // Set-level symmetric difference
-            let a = coerce_to_set(left);
-            let b = coerce_to_set(right);
-            Value::set(a.symmetric_difference(&b).cloned().collect())
-        }
-    }
-}
-
-/// Multi-arg symmetric difference: for each key, result = max_weight - second_max_weight.
-/// This is NOT a left-fold; it operates on all inputs simultaneously.
-pub(crate) fn set_sym_diff_multi(args: &[Value]) -> Value {
-    let level = args.iter().map(set_type_level).max().unwrap_or(0);
-    match level {
-        2 => {
-            // Mix-level: collect all weight vectors per key, then max - second_max
-            let maps: Vec<HashMap<String, f64>> = args.iter().map(to_mix_map).collect();
-            let mut all_keys: HashSet<String> = HashSet::new();
-            for m in &maps {
-                all_keys.extend(m.keys().cloned());
-            }
-            let mut result = HashMap::new();
-            for k in all_keys {
-                let mut weights: Vec<f64> = maps
-                    .iter()
-                    .map(|m| m.get(&k).copied().unwrap_or(0.0))
-                    .collect();
-                weights.sort_by(|a, b| b.partial_cmp(a).unwrap_or(std::cmp::Ordering::Equal));
-                let diff = weights[0] - weights.get(1).copied().unwrap_or(0.0);
-                if diff != 0.0 {
-                    result.insert(k, diff);
-                }
-            }
-            Value::mix(result)
-        }
-        1 => {
-            // Bag-level: collect all count vectors per key, then max - second_max
-            let maps: Vec<HashMap<String, i64>> = args.iter().map(to_bag_map).collect();
-            let mut all_keys: HashSet<String> = HashSet::new();
-            for m in &maps {
-                all_keys.extend(m.keys().cloned());
-            }
-            let mut result = HashMap::new();
-            for k in all_keys {
-                let mut counts: Vec<i64> = maps
-                    .iter()
-                    .map(|m| m.get(&k).copied().unwrap_or(0))
-                    .collect();
-                counts.sort_by(|a, b| b.cmp(a));
-                let diff = counts[0] - counts.get(1).copied().unwrap_or(0);
-                if diff > 0 {
-                    result.insert(k, diff);
-                }
-            }
-            Value::bag(result)
-        }
-        _ => {
-            // Set-level: element is in result iff it appears in exactly 1 input
-            let sets: Vec<HashSet<String>> = args.iter().map(coerce_to_set).collect();
-            let mut counts: HashMap<String, usize> = HashMap::new();
-            for s in &sets {
-                for k in s {
-                    *counts.entry(k.clone()).or_insert(0) += 1;
-                }
-            }
-            Value::set(
-                counts
-                    .into_iter()
-                    .filter(|(_, count)| *count == 1)
-                    .map(|(k, _)| k)
-                    .collect(),
-            )
-        }
-    }
-}
-
-/// Build an `X::Str::Numeric` error for a string that cannot be used as a
-/// number, carrying rakudo's `source`, `pos` and `reason` attributes.
-pub(crate) fn str_numeric_error(source: &str, pos: usize, reason: &str) -> RuntimeError {
-    let msg = format!(
-        "Cannot convert string to number: {} in '{}'",
-        reason, source
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("source".to_string(), Value::str(source.to_string()));
-    attrs.insert("pos".to_string(), Value::Int(pos as i64));
-    attrs.insert("reason".to_string(), Value::str(reason.to_string()));
-    attrs.insert("target-name".to_string(), Value::str("Numeric".to_string()));
-    attrs.insert(
-        "source-indicator".to_string(),
-        Value::str(crate::runtime::str_numeric::build_source_indicator(
-            source, pos,
-        )),
-    );
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(crate::symbol::Symbol::intern("X::Str::Numeric"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Raise `X::Str::Numeric` if `value` is a `Str` (or allomorph wrapping one)
-/// that cannot be parsed as a number. Used by the genuinely-numeric operators
-/// (`+ - * / % **`, `== != < > <= >= <=>`) so `"5 foo" + 8` fails the way Raku
-/// requires. The generic comparators (`cmp`, `before`/`after`, `min`/`max`) do
-/// NOT call this — they compare strings as strings.
-pub(crate) fn check_str_numeric(value: &Value) -> Result<(), RuntimeError> {
-    // Hot path: only a bare or Mixin-wrapped Str can fail; everything else
-    // (Int/Num/Rat/...) returns immediately without cloning or parsing.
-    let s = match value {
-        Value::Str(s) => s,
-        Value::Mixin(inner, _) => match inner.as_ref() {
-            Value::Str(s) => s,
-            _ => return Ok(()),
-        },
-        _ => return Ok(()),
-    };
-    if let Some((pos, reason)) = crate::runtime::str_numeric::str_numeric_failure(s) {
-        return Err(str_numeric_error(s, pos, &reason));
-    }
-    Ok(())
-}
-
-pub(crate) fn coerce_numeric(left: Value, right: Value) -> (Value, Value) {
-    // Unwrap allomorphic types (Mixin) to their inner numeric value
-    let left = unwrap_mixin(left);
-    let right = unwrap_mixin(right);
-    let l = match &left {
-        Value::Int(_)
-        | Value::BigInt(_)
-        | Value::Num(_)
-        | Value::Rat(_, _)
-        | Value::FatRat(_, _)
-        | Value::BigRat(_, _)
-        | Value::Complex(_, _) => left,
-        _ => coerce_to_numeric(left),
-    };
-    let r = match &right {
-        Value::Int(_)
-        | Value::BigInt(_)
-        | Value::Num(_)
-        | Value::Rat(_, _)
-        | Value::FatRat(_, _)
-        | Value::BigRat(_, _)
-        | Value::Complex(_, _) => right,
-        _ => coerce_to_numeric(right),
-    };
-    (l, r)
-}
-
-/// Unwrap a Mixin (allomorphic type) to its inner value.
-pub(crate) fn unwrap_mixin(val: Value) -> Value {
-    match val {
-        Value::Mixin(inner, _) => inner.as_ref().clone(),
-        other => other,
-    }
-}
-
-pub(crate) fn to_rat_parts(val: &Value) -> Option<(i64, i64)> {
-    match val {
-        Value::Mixin(inner, _) => to_rat_parts(inner),
-        Value::Int(i) => Some((*i, 1)),
-        Value::Rat(n, d) => Some((*n, *d)),
-        Value::FatRat(n, d) => Some((*n, *d)),
-        _ => None,
-    }
-}
-
-pub(crate) fn to_big_rat_parts(val: &Value) -> Option<(BigInt, BigInt)> {
-    match val {
-        Value::Mixin(inner, _) => to_big_rat_parts(inner),
-        Value::Int(i) => Some((BigInt::from(*i), BigInt::from(1))),
-        Value::BigInt(i) => Some(((**i).clone(), BigInt::from(1))),
-        Value::Rat(n, d) | Value::FatRat(n, d) => Some((BigInt::from(*n), BigInt::from(*d))),
-        Value::BigRat(n, d) => Some(((**n).clone(), (**d).clone())),
-        _ => None,
-    }
-}
-
-fn big_rat_parts_to_f64(num: &BigInt, den: &BigInt) -> f64 {
-    if den.is_zero() {
-        if num.is_zero() {
-            f64::NAN
-        } else if num.is_positive() {
-            f64::INFINITY
-        } else {
-            f64::NEG_INFINITY
-        }
-    } else {
-        num.to_f64().unwrap_or(0.0) / den.to_f64().unwrap_or(1.0)
-    }
-}
-
-pub(crate) fn compare_big_rat_parts(
-    a: (BigInt, BigInt),
-    b: (BigInt, BigInt),
-) -> Option<std::cmp::Ordering> {
-    let (an, ad) = a;
-    let (bn, bd) = b;
-    if ad.is_zero() || bd.is_zero() {
-        return big_rat_parts_to_f64(&an, &ad).partial_cmp(&big_rat_parts_to_f64(&bn, &bd));
-    }
-    Some((an * &bd).cmp(&(bn * &ad)))
-}
-
-pub(crate) fn big_rat_parts_equal(a: (BigInt, BigInt), b: (BigInt, BigInt)) -> bool {
-    let (an, ad) = a;
-    let (bn, bd) = b;
-    if ad.is_zero() || bd.is_zero() {
-        if (ad.is_zero() && an.is_zero()) || (bd.is_zero() && bn.is_zero()) {
-            return false;
-        }
-        return big_rat_parts_to_f64(&an, &ad) == big_rat_parts_to_f64(&bn, &bd);
-    }
-    let ga = an.gcd(&ad);
-    let gb = bn.gcd(&bd);
-    let mut an = an / &ga;
-    let mut ad = ad / ga;
-    let mut bn = bn / &gb;
-    let mut bd = bd / gb;
-    if ad.is_negative() {
-        an = -an;
-        ad = -ad;
-    }
-    if bd.is_negative() {
-        bn = -bn;
-        bd = -bd;
-    }
-    an == bn && ad == bd
-}
-
-fn rat_parts_to_f64(num: i64, den: i64) -> f64 {
-    if den != 0 {
-        num as f64 / den as f64
-    } else if num > 0 {
-        f64::INFINITY
-    } else if num < 0 {
-        f64::NEG_INFINITY
-    } else {
-        f64::NAN
-    }
-}
-
-pub(crate) fn compare_rat_parts(a: (i64, i64), b: (i64, i64)) -> std::cmp::Ordering {
-    let (an, ad) = a;
-    let (bn, bd) = b;
-    if ad == 0 || bd == 0 {
-        return rat_parts_to_f64(an, ad)
-            .partial_cmp(&rat_parts_to_f64(bn, bd))
-            .unwrap_or(std::cmp::Ordering::Equal);
-    }
-    let lhs = an as i128 * bd as i128;
-    let rhs = bn as i128 * ad as i128;
-    lhs.cmp(&rhs)
-}
-
-pub(crate) fn to_float_value(val: &Value) -> Option<f64> {
-    match val {
-        Value::ContainerRef(cell) => to_float_value(&cell.lock().unwrap()),
-        Value::Mixin(inner, _) => to_float_value(inner),
-        Value::Num(f) => Some(*f),
-        Value::Int(i) => Some(*i as f64),
-        Value::BigInt(n) => n.to_f64(),
-        Value::Rat(n, d) => {
-            if *d != 0 {
-                Some(*n as f64 / *d as f64)
-            } else if *n > 0 {
-                Some(f64::INFINITY)
-            } else if *n < 0 {
-                Some(f64::NEG_INFINITY)
-            } else {
-                Some(f64::NAN)
-            }
-        }
-        Value::FatRat(n, d) => {
-            if *d != 0 {
-                Some(*n as f64 / *d as f64)
-            } else if *n > 0 {
-                Some(f64::INFINITY)
-            } else if *n < 0 {
-                Some(f64::NEG_INFINITY)
-            } else {
-                Some(f64::NAN)
-            }
-        }
-        Value::BigRat(n, d) => {
-            let (n, d) = (n.as_ref(), d.as_ref());
-            if !d.is_zero() {
-                if let (Some(nn), Some(dd)) = (n.to_f64(), d.to_f64())
-                    && nn.is_finite()
-                    && dd.is_finite()
-                {
-                    Some(nn / dd)
-                } else {
-                    let scale_pow = 30u32;
-                    let scale = num_bigint::BigInt::from(10u8).pow(scale_pow);
-                    let scaled = (n * &scale) / d;
-                    if let Some(scaled_f) = scaled
-                        .to_f64()
-                        .or_else(|| scaled.to_string().parse::<f64>().ok())
-                    {
-                        Some(scaled_f / 10f64.powi(scale_pow as i32))
-                    } else if n.is_zero() {
-                        Some(0.0)
-                    } else if n.is_positive() {
-                        Some(f64::INFINITY)
-                    } else {
-                        Some(f64::NEG_INFINITY)
-                    }
-                }
-            } else if n.is_positive() {
-                Some(f64::INFINITY)
-            } else if n.is_negative() {
-                Some(f64::NEG_INFINITY)
-            } else {
-                Some(f64::NAN)
-            }
-        }
-        Value::Complex(r, i) => {
-            if *i == 0.0 {
-                Some(*r)
-            } else {
-                None
-            }
-        }
-        Value::Enum { value, .. } => Some(value.as_i64() as f64),
-        Value::Bool(b) => Some(if *b { 1.0 } else { 0.0 }),
-        Value::Str(s) => s.trim().parse::<f64>().ok(),
-        Value::Nil => Some(0.0),
-        Value::Set(items, _) => Some(items.len() as f64),
-        Value::Bag(items, _) => Some(items.len() as f64),
-        Value::Mix(items, _) => Some(items.len() as f64),
-        Value::Hash(items) => Some(items.len() as f64),
-        _ if val.as_list_items().is_some() => Some(val.as_list_items().unwrap().len() as f64),
-        Value::LazyList(ll) => {
-            if let Some(cached) = ll.cache.lock().unwrap().as_ref() {
-                Some(cached.len() as f64)
-            } else {
-                Some(0.0)
-            }
-        }
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Instant" => attributes.as_map().get("value").and_then(to_float_value),
-        Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } if class_name == "Match" => attributes.as_map().get("str").and_then(to_float_value),
-        Value::Instance { attributes, .. } => attributes
-            .as_map()
-            .get("__mutsu_int_value")
-            .and_then(to_float_value),
-        // A type object (e.g. `Any`, `Str`, `Rat`, a user class) numifies to 0 in
-        // numeric context (Raku warns "Use of uninitialized value"). This makes
-        // `(Any) == 0` / `@a[oob] == 0` true, matching Rakudo. (`Mu` alone has no
-        // Numeric and dies in raku, but treating it as 0 is a harmless divergence.)
-        Value::Package(_) => Some(0.0),
-        _ => None,
-    }
-}
-
-pub(crate) fn to_complex_parts(val: &Value) -> Option<(f64, f64)> {
-    match val {
-        Value::Complex(r, i) => Some((*r, *i)),
-        _ => to_float_value(val).map(|v| (v, 0.0)),
-    }
-}
-
-pub(crate) fn compare_values(a: &Value, b: &Value) -> i32 {
-    fn compare_infinite_num_against_nonnumeric_str(num: f64, s: &str) -> Option<i32> {
-        if !num.is_infinite() || s.trim().parse::<f64>().is_ok() {
-            return None;
-        }
-        Some(if num.is_sign_positive() {
-            std::cmp::Ordering::Greater
-        } else {
-            std::cmp::Ordering::Less
-        } as i32)
-    }
-
-    // First-class element containers (`ContainerRef`, e.g. a `:=`-bound array
-    // slot or a `.grep` rw alias) compare by their inner value — `.sort`/min/max
-    // over a grepped/bound array must order the values, not the cells.
-    if matches!(a, Value::ContainerRef(_)) || matches!(b, Value::ContainerRef(_)) {
-        return compare_values(&a.deref_container(), &b.deref_container());
-    }
-    match (a, b) {
-        (Value::Version { parts: ap, .. }, Value::Version { parts: bp, .. }) => {
-            version_cmp_parts(ap, bp) as i32
-        }
-        (Value::Int(a), Value::Int(b)) => a.cmp(b) as i32,
-        (Value::BigInt(a), Value::BigInt(b)) => a.as_ref().cmp(b.as_ref()) as i32,
-        (Value::BigInt(a), Value::Int(b)) => a.as_ref().cmp(&num_bigint::BigInt::from(*b)) as i32,
-        (Value::Int(a), Value::BigInt(b)) => num_bigint::BigInt::from(*a).cmp(b.as_ref()) as i32,
-        (Value::Num(a), Value::Num(b)) => {
-            // NaN sorts after everything (including Inf)
-            match (a.is_nan(), b.is_nan()) {
-                (true, true) => 0,
-                (true, false) => 1,
-                (false, true) => -1,
-                _ => a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal) as i32,
-            }
-        }
-        (Value::BigInt(a), Value::Num(b)) => {
-            a.as_ref()
-                .to_f64()
-                .unwrap_or(if a.as_ref().is_positive() {
-                    f64::INFINITY
-                } else {
-                    f64::NEG_INFINITY
-                })
-                .partial_cmp(b)
-                .unwrap_or(std::cmp::Ordering::Equal) as i32
-        }
-        (Value::Num(a), Value::BigInt(b)) => {
-            a.partial_cmp(&b.as_ref().to_f64().unwrap_or(if b.as_ref().is_positive() {
-                f64::INFINITY
-            } else {
-                f64::NEG_INFINITY
-            }))
-            .unwrap_or(std::cmp::Ordering::Equal) as i32
-        }
-        (Value::Int(a), Value::Num(b)) => (*a as f64)
-            .partial_cmp(b)
-            .unwrap_or(std::cmp::Ordering::Equal) as i32,
-        (Value::Num(a), Value::Int(b)) => a
-            .partial_cmp(&(*b as f64))
-            .unwrap_or(std::cmp::Ordering::Equal) as i32,
-        (Value::Num(a), Value::Rat(n, d)) => {
-            let rat_f = if *d != 0 {
-                *n as f64 / *d as f64
-            } else {
-                f64::NAN
-            };
-            a.partial_cmp(&rat_f).unwrap_or(std::cmp::Ordering::Equal) as i32
-        }
-        (Value::Rat(n, d), Value::Num(b)) => {
-            let rat_f = if *d != 0 {
-                *n as f64 / *d as f64
-            } else {
-                f64::NAN
-            };
-            rat_f.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal) as i32
-        }
-        (Value::Num(n), Value::Str(s)) => {
-            if let Some(ord) = compare_infinite_num_against_nonnumeric_str(*n, s) {
-                ord
-            } else {
-                a.to_string_value().cmp(&b.to_string_value()) as i32
-            }
-        }
-        (Value::Str(s), Value::Num(n)) => {
-            if let Some(ord) = compare_infinite_num_against_nonnumeric_str(*n, s) {
-                -ord
-            } else {
-                a.to_string_value().cmp(&b.to_string_value()) as i32
-            }
-        }
-        // Pair/ValuePair comparison: compare by key first, then by value
-        (Value::Pair(ak, av), Value::Pair(bk, bv)) => {
-            let key_cmp = compare_values(&Value::str(ak.clone()), &Value::str(bk.clone()));
-            if key_cmp != 0 {
-                key_cmp
-            } else {
-                compare_values(av, bv)
-            }
-        }
-        (Value::ValuePair(ak, av), Value::ValuePair(bk, bv)) => {
-            let key_cmp = compare_values(ak, bk);
-            if key_cmp != 0 {
-                key_cmp
-            } else {
-                compare_values(av, bv)
-            }
-        }
-        (Value::Pair(ak, av), Value::ValuePair(bk, bv)) => {
-            let key_cmp = compare_values(&Value::str(ak.clone()), bk);
-            if key_cmp != 0 {
-                key_cmp
-            } else {
-                compare_values(av, bv)
-            }
-        }
-        (Value::ValuePair(ak, av), Value::Pair(bk, bv)) => {
-            let key_cmp = compare_values(ak, &Value::str(bk.clone()));
-            if key_cmp != 0 {
-                key_cmp
-            } else {
-                compare_values(av, bv)
-            }
-        }
-        // Enum values: compare by their integer value
-        (Value::Enum { value: av, .. }, Value::Enum { value: bv, .. }) => {
-            av.as_i64().cmp(&bv.as_i64()) as i32
-        }
-        _ => {
-            if let (Some((an, ad)), Some((bn, bd))) = (to_rat_parts(a), to_rat_parts(b)) {
-                let cmp = compare_rat_parts((an, ad), (bn, bd)) as i32;
-                if cmp != 0 {
-                    return cmp;
-                }
-                // For allomorphic types (IntStr, RatStr, etc.), break numeric ties
-                // with string comparison
-                if matches!(a, Value::Mixin(..)) || matches!(b, Value::Mixin(..)) {
-                    return a.to_string_value().cmp(&b.to_string_value()) as i32;
-                }
-                return cmp;
-            }
-            a.to_string_value().cmp(&b.to_string_value()) as i32
-        }
-    }
-}
-
-pub(crate) fn to_int(v: &Value) -> i64 {
-    match v {
-        // Phase 2 element container: a `:=`-bound element cell that leaked into
-        // a numeric context reads through to its inner value.
-        Value::ContainerRef(cell) => to_int(&cell.lock().unwrap()),
-        Value::Mixin(inner, _) => to_int(inner),
-        Value::Int(i) => *i,
-        Value::BigInt(n) => {
-            use num_traits::ToPrimitive;
-            n.as_ref()
-                .to_i64()
-                .unwrap_or(if **n > num_bigint::BigInt::from(0i64) {
-                    i64::MAX
-                } else {
-                    i64::MIN
-                })
-        }
-        Value::Num(f) => *f as i64,
-        Value::Range(a, b) => {
-            if b >= a {
-                b - a + 1
-            } else {
-                0
-            }
-        }
-        Value::RangeExcl(a, b) | Value::RangeExclStart(a, b) => {
-            if b > a {
-                b - a
-            } else {
-                0
-            }
-        }
-        Value::RangeExclBoth(a, b) => {
-            if *b > *a + 1 {
-                b - a - 1
-            } else {
-                0
-            }
-        }
-        Value::Rat(n, d) => {
-            if *d != 0 {
-                n / d
-            } else {
-                0
-            }
-        }
-        Value::Complex(r, _) => *r as i64,
-        Value::Str(s) => s.parse().unwrap_or(0),
-        Value::Array(items, ..) => items.len() as i64,
-        Value::Hash(items) => items.len() as i64,
-        Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) => items.len() as i64,
-        Value::Slip(items) => items.len() as i64,
-        Value::Capture { positional, .. } => positional.len() as i64,
-        Value::Instance { attributes, .. } => attributes
-            .as_map()
-            .get("__mutsu_int_value")
-            .map_or(0, to_int),
-        _ => 0,
-    }
-}
-
-pub(crate) fn merge_junction(kind: JunctionKind, left: Value, right: Value) -> Value {
-    // Infix junction operators (|, &, ^) always create a new 2-element
-    // junction without flattening. List-associative flattening is handled
-    // at compile time via JunctionAnyN/AllN/OneN opcodes.
-    Value::junction(kind, vec![left, right])
-}
-
-/// Format a short representation of a value for type-check error messages,
-/// matching Raku's format: e.g. `("hello")`, `(42)`.
-pub(crate) fn value_short_repr(val: &Value) -> String {
-    match val {
-        Value::Str(s) => format!("(\"{}\")", s),
-        Value::Int(n) => format!("({})", n),
-        Value::BigInt(n) => format!("({})", n),
-        Value::Num(n) => format!("({})", n),
-        Value::Bool(b) => format!("({})", if *b { "True" } else { "False" }),
-        Value::Rat(n, d) => format!("({}/{})", n, d),
-        Value::BigRat(n, d) => format!("({}/{})", n, d),
-        Value::FatRat(n, d) => format!("(FatRat.new({}, {}))", n, d),
-        Value::Nil => "(Nil)".to_string(),
-        _ => String::new(),
-    }
-}
-
-/// Format the variable name for error messages, adding `$` sigil for
-/// scalar variables that don't already have a sigil prefix.
-pub(crate) fn format_var_name_for_error(name: &str) -> String {
-    if name.starts_with('$')
-        || name.starts_with('@')
-        || name.starts_with('%')
-        || name.starts_with('&')
-    {
-        name.to_string()
-    } else {
-        format!("${}", name)
-    }
-}
-
-/// Build the standard X::TypeCheck::Assignment error message, matching Raku's format:
-/// `Type check failed in assignment to $x; expected Int but got Str ("hello")`
-pub(crate) fn type_check_assignment_error(var_name: &str, expected: &str, val: &Value) -> String {
-    let display_name = format_var_name_for_error(var_name);
-    let got_type = value_type_name(val);
-    let repr = value_short_repr(val);
-    if repr.is_empty() {
-        format!(
-            "X::TypeCheck::Assignment: Type check failed in assignment to {}; expected {} but got {}",
-            display_name, expected, got_type
-        )
-    } else {
-        format!(
-            "X::TypeCheck::Assignment: Type check failed in assignment to {}; expected {} but got {} {}",
-            display_name, expected, got_type, repr
-        )
-    }
-}
-
-/// Build a structured X::TypeCheck::Binding RuntimeError, for `:=` binds to a
-/// typed scalar (e.g. `my Str $x := 3`), matching Raku's format:
-/// `Type check failed in binding; expected Str but got Int (3)`
-pub(crate) fn type_check_binding_typed_error(expected: &str, val: &Value) -> RuntimeError {
-    let got_type = value_type_name(val);
-    let repr = value_short_repr(val);
-    let msg = if repr.is_empty() {
-        format!(
-            "X::TypeCheck::Binding: Type check failed in binding; expected {} but got {}",
-            expected, got_type
-        )
-    } else {
-        format!(
-            "X::TypeCheck::Binding: Type check failed in binding; expected {} but got {} {}",
-            expected, got_type, repr
-        )
-    };
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert(
-        "expected".to_string(),
-        crate::value::expected_type_object(expected),
-    );
-    attrs.insert("got".to_string(), val.clone());
-    attrs.insert("operation".to_string(), Value::str("bind".to_string()));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    RuntimeError::typed("X::TypeCheck::Binding", attrs)
-}
-
-/// Build a structured X::Dynamic::NotFound RuntimeError.
-/// Thrown when assigning to a dynamic variable (`$*x` / `@*x` / `%*x`) that is
-/// not present anywhere in the dynamic scope. `display_name` is the full
-/// sigil+twigil form (e.g. `$*an_undeclared_dynvar`).
-pub(crate) fn dynamic_not_found_error(display_name: &str) -> RuntimeError {
-    let msg = format!("Dynamic variable {} not found", display_name);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("name".to_string(), Value::str(display_name.to_string()));
-    attrs.insert("symbol".to_string(), Value::str(display_name.to_string()));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    RuntimeError::typed("X::Dynamic::NotFound", attrs)
-}
-
-/// Build a structured X::Caller::NotDynamic RuntimeError.
-/// Thrown when accessing a caller-frame lexical through `CALLER::` (either the
-/// `$CALLER::x` symbolic form or the `CALLER::<$x>` stash-subscript form) when
-/// that variable is not declared `is dynamic`. `name` is the bare variable name
-/// without sigil; the reported `symbol` re-adds the `$` sigil.
-pub(crate) fn caller_not_dynamic_error(name: &str) -> RuntimeError {
-    let symbol = format!("${name}");
-    let msg =
-        format!("Cannot access '{symbol}' through CALLER, because it is not declared as dynamic");
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("symbol".to_string(), Value::str(symbol));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    RuntimeError::typed("X::Caller::NotDynamic", attrs)
-}
-
-/// Build a structured X::TypeCheck::Assignment RuntimeError.
-/// This creates a proper exception object that `throws-like` can match.
-pub(crate) fn type_check_assignment_typed_error(
-    var_name: &str,
-    expected: &str,
-    val: &Value,
-) -> RuntimeError {
-    let msg = type_check_assignment_error(var_name, expected, val);
-    let display_name = format_var_name_for_error(var_name);
-    let mut attrs = std::collections::HashMap::new();
-    // raku exposes `.expected` as the expected type OBJECT and `.got` as the
-    // offending VALUE (not its type name), so `throws-like` matchers like
-    // `expected => Int` / `got => 'foo'` succeed.
-    attrs.insert(
-        "expected".to_string(),
-        crate::value::expected_type_object(expected),
-    );
-    attrs.insert("got".to_string(), val.clone());
-    attrs.insert("symbol".to_string(), Value::str(display_name));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    RuntimeError::typed("X::TypeCheck::Assignment", attrs)
-}
-
-/// Build the standard X::TypeCheck::Assignment error for array/hash elements:
-/// `Type check failed for an element of @a; expected Int but got Str ("hi")`
-pub(crate) fn type_check_element_error(var_name: &str, expected: &str, val: &Value) -> String {
-    let display_name = format_var_name_for_error(var_name);
-    let got_type = value_type_name(val);
-    let repr = value_short_repr(val);
-    if repr.is_empty() {
-        format!(
-            "X::TypeCheck::Assignment: Type check failed for an element of {}; expected {} but got {}",
-            display_name, expected, got_type
-        )
-    } else {
-        format!(
-            "X::TypeCheck::Assignment: Type check failed for an element of {}; expected {} but got {} {}",
-            display_name, expected, got_type, repr
-        )
-    }
-}
-
-/// Build a structured X::TypeCheck::Assignment RuntimeError for element type checks.
-pub(crate) fn type_check_element_typed_error(
-    var_name: &str,
-    expected: &str,
-    val: &Value,
-) -> RuntimeError {
-    let msg = type_check_element_error(var_name, expected, val);
-    let display_name = format_var_name_for_error(var_name);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("expected".to_string(), Value::str(expected.to_string()));
-    // `.got` is the offending value itself (e.g. `42`), so `$!.got ~~ Int` holds,
-    // matching Rakudo's X::TypeCheck.
-    attrs.insert("got".to_string(), val.clone());
-    attrs.insert("symbol".to_string(), Value::str(display_name));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    RuntimeError::typed("X::TypeCheck::Assignment", attrs)
-}
-
-/// Compute the `.WHICH` string for a value, used as the internal key
-/// in object hashes (`my %h{Any}`).
-pub(crate) fn value_which_key(value: &Value) -> String {
-    match value {
-        Value::Int(n) => format!("Int|{}", n),
-        Value::BigInt(n) => format!("Int|{}", n),
-        Value::Num(n) => format!("Num|{}", n),
-        Value::Str(s) => format!("Str|{}", s),
-        Value::Bool(b) => format!("Bool|{}", if *b { 1 } else { 0 }),
-        Value::Rat(n, d) => format!("Rat|{}/{}", n, d),
-        Value::FatRat(n, d) => format!("FatRat|{}/{}", n, d),
-        Value::BigRat(n, d) => format!("Rat|{}/{}", n, d),
-        Value::Complex(r, i) => format!("Complex|{}+{}i", r, i),
-        Value::Nil => format!("Nil|U{}", Symbol::intern("Nil").id()),
-        Value::Package(name) => format!("{}|U{}", name.resolve(), name.id()),
-        Value::CustomType(c) => format!("{}|U{}", c.name.resolve(), c.id),
-        Value::Instance { id, .. } => {
-            format!("{}|{}", value_type_name(value), id)
-        }
-        Value::Array(items, ..) => format!("Array|{:p}", Arc::as_ptr(items)),
-        Value::Hash(map) => format!("Hash|{:p}", Arc::as_ptr(map)),
-        Value::Pair(k, v) => format!("Pair|{}|{}", k, value_which_key(v)),
-        Value::ValuePair(k, v) => format!("Pair|{}|{}", value_which_key(k), value_which_key(v)),
-        Value::Enum { enum_type, key, .. } => {
-            format!("{}|{}", enum_type.resolve(), key.resolve())
-        }
-        // A `but`-mixed value (e.g. `"quux" but $role`) keeps the identity of
-        // its base value but is a DISTINCT object per mixed-in role set — raku's
-        // `.WHICH` is `Str+{<role>}|quux`. Fold the (sorted) mixin type/role
-        // names into the key so two different roles over the same base value are
-        // distinct object-hash keys, while the same value+role collides.
-        Value::Mixin(inner, mixins) => {
-            let mut roles: Vec<&str> = mixins.keys().map(|s| s.as_str()).collect();
-            roles.sort_unstable();
-            format!(
-                "{}+{{{}}}|{}",
-                value_type_name(inner),
-                roles.join(","),
-                value_which_key(inner)
-            )
-        }
-        _ => {
-            format!("{}|{}", value_type_name(value), value.to_string_value())
-        }
-    }
-}
