+++ src/runtime/utils/type_constraints.rs
@@ -1,0 +1,479 @@
+pub(crate) fn is_known_type_constraint(constraint: &str) -> bool {
+    if crate::runtime::native_types::is_native_int_type(constraint) {
+        return true;
+    }
+    // Handle parameterized types like Array[Int], Hash[Str], etc.
+    if let Some(base) = constraint.split('[').next()
+        && constraint.contains('[')
+        && constraint.ends_with(']')
+        && is_known_type_constraint(base)
+    {
+        return true;
+    }
+    matches!(
+        constraint,
+        "Int"
+            | "UInt"
+            | "Num"
+            | "Str"
+            | "Bool"
+            | "Array"
+            | "Hash"
+            | "Rat"
+            | "FatRat"
+            | "Complex"
+            | "atomicint"
+            | "int"
+            | "num"
+            | "num32"
+            | "num64"
+            | "str"
+            // Core types
+            | "Mu"
+            | "Any"
+            | "Cool"
+            | "Nil"
+            | "Pair"
+            | "Map"
+            | "List"
+            | "Slip"
+            | "Seq"
+            | "Range"
+            | "Set"
+            | "SetHash"
+            | "Bag"
+            | "BagHash"
+            | "Mix"
+            | "MixHash"
+            | "QuantHash"
+            | "Blob"
+            | "Buf"
+            // Native sized buffer aliases: blob8 == Blob[uint8], buf16 == Buf[uint16], etc.
+            // (utf8/utf16/utf32 are already listed below.)
+            | "blob"
+            | "buf"
+            | "blob8"
+            | "blob16"
+            | "blob32"
+            | "blob64"
+            | "buf8"
+            | "buf16"
+            | "buf32"
+            | "buf64"
+            | "Junction"
+            | "Match"
+            | "Regex"
+            | "Block"
+            | "Code"
+            | "Sub"
+            | "Method"
+            | "Submethod"
+            | "Routine"
+            | "Capture"
+            | "Signature"
+            | "Parameter"
+            | "Whatever"
+            | "WhateverCode"
+            | "HyperWhatever"
+            | "Format"
+            | "Version"
+            | "Date"
+            | "DateTime"
+            | "Instant"
+            | "Duration"
+            | "Failure"
+            | "Exception"
+            | "Backtrace"
+            | "Scalar"
+            | "Proxy"
+            | "Label"
+            | "Grammar"
+            // Allomorphs
+            | "Allomorph"
+            | "IntStr"
+            | "NumStr"
+            | "RatStr"
+            | "ComplexStr"
+            // Meta types
+            | "Attribute"
+            | "AST"
+            | "Compiler"
+            | "Distro"
+            | "Kernel"
+            | "Macro"
+            | "ObjAt"
+            | "ValueObjAt"
+            | "Stash"
+            | "PseudoStash"
+            | "Raku"
+            | "Perl"
+            | "VM"
+            | "Variable"
+            // Array-like
+            | "array"
+            | "Hashray"
+            // Numeric subtypes
+            | "byte"
+            | "uint"
+            // IO types
+            | "IO"
+            // Ref types
+            | "IntAttrRef"
+            | "IntLexRef"
+            | "IntPosRef"
+            | "NumAttrRef"
+            | "NumLexRef"
+            | "NumPosRef"
+            | "StrAttrRef"
+            | "StrLexRef"
+            | "StrPosRef"
+            | "UIntAttrRef"
+            | "UIntLexRef"
+            | "UIntPosRef"
+            | "ScalarVAR"
+            // String types
+            | "StrDistance"
+            // Iteration
+            | "IterationBuffer"
+            // Deprecation
+            | "Deprecation"
+            | "Distribution"
+            // Operators
+            | "Operator"
+            | "OperatorProperties"
+            // Unicode types
+            | "Unicode"
+            | "Uni"
+            | "NFC"
+            | "NFD"
+            | "NFKC"
+            | "NFKD"
+            | "utf8"
+            | "utf16"
+            | "utf32"
+            | "SignedBlob"
+            | "UnsignedBlob"
+            // Concurrent types
+            | "Cancellation"
+            | "Channel"
+            | "CurrentThreadScheduler"
+            | "Promise"
+            | "Semaphore"
+            | "Supplier"
+            | "Tap"
+            | "Thread"
+            | "ThreadPoolScheduler"
+            // Hyper/Parallel
+            | "HyperConfiguration"
+            | "HyperSeq"
+            | "ParallelSequence"
+            // Slang
+            | "Slang"
+            // Lock
+            | "Lock"
+            // CallFrame
+            | "CallFrame"
+            // CompUnit
+            | "CompUnit"
+    )
+}
+
+/// Check if a compound (`::`-separated) name is a known Raku standard type.
+pub(crate) fn is_known_compound_type(name: &str) -> bool {
+    matches!(
+        name,
+        "Backtrace::Frame"
+            | "CompUnit::DependencySpecification"
+            | "CompUnit::RepositoryRegistry"
+            | "CompUnit::Repository::FileSystem"
+            | "CompUnit::Repository::Installable"
+            | "CompUnit::Repository::Installation"
+            | "CompUnit::Repository::Locally"
+            | "CompUnit::Repository::NQP"
+            | "CompUnit::Repository::Perl5"
+            | "Distribution"
+            | "Distribution::Hash"
+            | "Distribution::Installation"
+            | "Distribution::Path"
+            | "IO::ArgFiles"
+            | "IO::CatHandle"
+            | "IO::Handle"
+            | "IO::Notification"
+            | "IO::Path"
+            | "IO::Path::Cygwin"
+            | "IO::Path::QNX"
+            | "IO::Path::Unix"
+            | "IO::Path::Win32"
+            | "IO::Pipe"
+            | "IO::Socket::INET"
+            | "IO::Socket::Async"
+            | "IO::Spec"
+            | "IO::Spec::Cygwin"
+            | "IO::Spec::QNX"
+            | "IO::Spec::Unix"
+            | "IO::Spec::Win32"
+            | "IO::Special"
+            | "Metamodel::Primitives"
+            | "Pod::Block"
+            | "Pod::Block::Code"
+            | "Pod::Block::Comment"
+            | "Pod::Block::Declarator"
+            | "Pod::Block::Named"
+            | "Pod::Block::Para"
+            | "Pod::Block::Table"
+            | "Pod::Config"
+            | "Pod::Defn"
+            | "Pod::FormattingCode"
+            | "Pod::Heading"
+            | "Pod::Item"
+            | "Pod::Raw"
+            | "Proc::Async"
+            // Exception types
+            | "X::AdHoc"
+            | "X::Adverb"
+            | "X::Anon::Augment"
+            | "X::Anon::Multi"
+            | "X::Assignment::RO"
+            | "X::Assignment::RO::Comp"
+            | "X::Attribute::NoPackage"
+            | "X::Attribute::Package"
+            | "X::Attribute::Regex"
+            | "X::Attribute::Required"
+            | "X::Attribute::Undeclared"
+            | "X::Augment::NoSuchType"
+            | "X::Backslash::NonVariableDollar"
+            | "X::Backslash::UnrecognizedSequence"
+            | "X::Bind"
+            | "X::Bind::NativeType"
+            | "X::Bind::Slice"
+            | "X::Bind::ZenSlice"
+            | "X::Buf::AsStr"
+            | "X::Buf::Pack"
+            | "X::Buf::Pack::NonASCII"
+            | "X::Caller::NotDynamic"
+            | "X::Cannot::Empty"
+            | "X::Cannot::Lazy"
+            | "X::Cannot::New"
+            | "X::Channel::ReceiveOnClosed"
+            | "X::Channel::SendOnClosed"
+            | "X::Comp::AdHoc"
+            | "X::Comp::BeginTime"
+            | "X::Comp::Group"
+            | "X::Comp::NYI"
+            | "X::Comp::Trait::NotOnNative"
+            | "X::Comp::Trait::Scope"
+            | "X::Comp::Trait::Unknown"
+            | "X::Composition::NotComposable"
+            | "X::Constructor::Positional"
+            | "X::ControlFlow"
+            | "X::ControlFlow::Return"
+            | "X::DateTime::InvalidDeltaUnit"
+            | "X::DateTime::TimezoneClash"
+            | "X::Declaration::OurScopeInRole"
+            | "X::Declaration::Scope"
+            | "X::Declaration::Scope::Multi"
+            | "X::Does::TypeObject"
+            | "X::Dynamic::NotFound"
+            | "X::Dynamic::Package"
+            | "X::Dynamic::Postdeclaration"
+            | "X::Eval::NoSuchLang"
+            | "X::Export::NameClash"
+            | "X::EXPORTHOW::Conflict"
+            | "X::EXPORTHOW::InvalidDirective"
+            | "X::EXPORTHOW::NothingToSupersede"
+            | "X::Hash::Store::OddNumber"
+            | "X::HyperOp::Infinite"
+            | "X::HyperOp::NonDWIM"
+            | "X::HyperWhatever::Multiple"
+            | "X::IllegalOnFixedDimensionArray"
+            | "X::IO::Chdir"
+            | "X::IO::Chmod"
+            | "X::IO::Copy"
+            | "X::IO::Cwd"
+            | "X::IO::Dir"
+            | "X::IO::Directory"
+            | "X::IO::DoesNotExist"
+            | "X::IO::Link"
+            | "X::IO::Mkdir"
+            | "X::IO::NotAChild"
+            | "X::IO::NotAFile"
+            | "X::IO::Resolve"
+            | "X::IO::Rename"
+            | "X::IO::Rmdir"
+            | "X::IO::Symlink"
+            | "X::IO::Unlink"
+            | "X::Immutable"
+            | "X::Import::MissingSymbols"
+            | "X::Import::NoSuchTag"
+            | "X::Import::OnlystarProto"
+            | "X::Import::Positional"
+            | "X::Import::Redeclaration"
+            | "X::Inheritance::NotComposed"
+            | "X::Inheritance::SelfInherit"
+            | "X::Inheritance::UnknownParent"
+            | "X::Inheritance::Unsupported"
+            | "X::InvalidType"
+            | "X::InvalidTypeSmiley"
+            | "X::Item"
+            | "X::Localizer::NoContainer"
+            | "X::Lock::ConditionVariable::New"
+            | "X::Match::Bool"
+            | "X::Method::InvalidQualifier"
+            | "X::Method::NotFound"
+            | "X::Method::Private::Permission"
+            | "X::Method::Private::Unqualified"
+            | "X::Mixin::NotComposable"
+            | "X::Multi::Ambiguous"
+            | "X::Multi::NoMatch"
+            | "X::NQP::NotFound"
+            | "X::NYI"
+            | "X::NYI::Available"
+            | "X::NYI::BigInt"
+            | "X::NoDispatcher"
+            | "X::NoSuchSymbol"
+            | "X::NotEnoughDimensions"
+            | "X::NotParametric"
+            | "X::Numeric::DivideByZero"
+            | "X::Numeric::Real"
+            | "X::Numeric::Underflow"
+            | "X::Obsolete"
+            | "X::OutOfRange"
+            | "X::Package::Stubbed"
+            | "X::Package::UseLib"
+            | "X::Pairup::OddNumber"
+            | "X::Parameter::AfterDefault"
+            | "X::Parameter::BadType"
+            | "X::Parameter::Default"
+            | "X::Parameter::Default::TypeCheck"
+            | "X::Parameter::InvalidType"
+            | "X::Parameter::MultipleTypeConstraints"
+            | "X::Parameter::Placeholder"
+            | "X::Parameter::RW"
+            | "X::Parameter::Twigil"
+            | "X::Parameter::TypedSlurpy"
+            | "X::Parameter::WrongOrder"
+            | "X::Phaser::Multiple"
+            | "X::Phaser::PrePost"
+            | "X::PhaserExceptions"
+            | "X::Placeholder::Attribute"
+            | "X::Placeholder::Block"
+            | "X::Placeholder::Mainline"
+            | "X::Placeholder::NonPlaceholder"
+            | "X::PoisonedAlias"
+            | "X::Pragma::CannotPrecomp"
+            | "X::Pragma::CannotWhat"
+            | "X::Pragma::MustOneOf"
+            | "X::Pragma::NoArgs"
+            | "X::Pragma::OnlyOne"
+            | "X::Pragma::UnknownArg"
+            | "X::Proc::Async::AlreadyStarted"
+            | "X::Proc::Async::CharsOrBytes"
+            | "X::Proc::Async::MustBeStarted"
+            | "X::Proc::Async::OpenForWriting"
+            | "X::Proc::Async::TapBeforeSpawn"
+            | "X::Proc::Unsuccessful"
+            | "X::Promise::CauseOnlyValidOnBroken"
+            | "X::Promise::Combinator"
+            | "X::Promise::Vowed"
+            | "X::PseudoPackage::InDeclaration"
+            | "X::Range::InvalidArg"
+            | "X::Redeclaration"
+            | "X::Redeclaration::Outer"
+            | "X::Role::Initialization"
+            | "X::Role::Parametric::NoSuchCandidate"
+            | "X::Routine::Unwrap"
+            | "X::Seq::Consumed"
+            | "X::Seq::NotIndexable"
+            | "X::Sequence::Deduction"
+            | "X::Set::Coerce"
+            | "X::Signature::NameClash"
+            | "X::Signature::Placeholder"
+            | "X::Str::Match::x"
+            | "X::Str::Numeric"
+            | "X::Str::Trans::IllegalKey"
+            | "X::Str::Trans::InvalidArg"
+            | "X::StubCode"
+            | "X::Subscript::Negative"
+            | "X::Supply::Migrate::Needs"
+            | "X::Syntax::AddCategorical::TooFewParts"
+            | "X::Syntax::AddCategorical::TooManyParts"
+            | "X::Syntax::Adverb"
+            | "X::Syntax::Argument::MOPMacro"
+            | "X::Syntax::Augment::Illegal"
+            | "X::Syntax::Augment::WithoutMonkeyTyping"
+            | "X::Syntax::BlockGobbled"
+            | "X::Syntax::CannotMeta"
+            | "X::Syntax::Comment::Embedded"
+            | "X::Syntax::ConditionalOperator::PrecedenceTooLoose"
+            | "X::Syntax::ConditionalOperator::SecondPartGobbled"
+            | "X::Syntax::ConditionalOperator::SecondPartInvalid"
+            | "X::Syntax::Confused"
+            | "X::Syntax::DuplicatedPrefix"
+            | "X::Syntax::Extension::Category"
+            | "X::Syntax::Extension::Null"
+            | "X::Syntax::Extension::SpecialForm"
+            | "X::Syntax::Extension::TooComplex"
+            | "X::Syntax::InfixInTermPosition"
+            | "X::Syntax::KeywordAsFunction"
+            | "X::Syntax::Malformed"
+            | "X::Syntax::Malformed::Elsif"
+            | "X::Syntax::Missing"
+            | "X::Syntax::Name::Null"
+            | "X::Syntax::NegatedPair"
+            | "X::Syntax::NoSelf"
+            | "X::Syntax::NonAssociative"
+            | "X::Syntax::Number::IllegalDecimal"
+            | "X::Syntax::Number::RadixOutOfRange"
+            | "X::Syntax::P5"
+            | "X::Syntax::Perl5Var"
+            | "X::Syntax::Pod::BeginWithoutEnd"
+            | "X::Syntax::Pod::BeginWithoutIdentifier"
+            | "X::Syntax::Regex::Adverb"
+            | "X::Syntax::Regex::MalformedRange"
+            | "X::Syntax::Regex::NonQuantifiable"
+            | "X::Syntax::Regex::QuantifierValue"
+            | "X::Syntax::Regex::NullRegex"
+            | "X::Syntax::Regex::SolitaryBacktrackControl"
+            | "X::Syntax::Regex::SolitaryQuantifier"
+            | "X::Syntax::Regex::SpacesInBareRange"
+            | "X::Syntax::Regex::UnrecognizedMetachar"
+            | "X::Syntax::Regex::UnrecognizedModifier"
+            | "X::Syntax::Regex::Unspace"
+            | "X::Syntax::Regex::Unterminated"
+            | "X::Syntax::Reserved"
+            | "X::Syntax::Self::WithoutObject"
+            | "X::Syntax::Signature::InvocantMarker"
+            | "X::Syntax::Term::MissingInitializer"
+            | "X::Syntax::UnlessElse"
+            | "X::Syntax::Variable::BadType"
+            | "X::Syntax::Variable::ConflictingTypes"
+            | "X::Syntax::Variable::IndirectDeclaration"
+            | "X::Syntax::Variable::Initializer"
+            | "X::Syntax::Variable::Match"
+            | "X::Syntax::Variable::MissingInitializer"
+            | "X::Syntax::Variable::Numeric"
+            | "X::Syntax::Variable::Twigil"
+            | "X::Syntax::VirtualCall"
+            | "X::Temporal::InvalidFormat"
+            | "X::TooLateForREPR"
+            | "X::Trait::NotOnNative"
+            | "X::Trait::Scope"
+            | "X::Trait::Unknown"
+            | "X::TypeCheck"
+            | "X::TypeCheck::Argument"
+            | "X::TypeCheck::Assignment"
+            | "X::TypeCheck::Binding"
+            | "X::TypeCheck::Return"
+            | "X::TypeCheck::Splice"
+            | "X::Undeclared"
+            | "X::Undeclared::Symbols"
+            | "X::UnitScope::Invalid"
+            | "X::UnitScope::TooLate"
+            | "X::Value::Dynamic"
+            | "X::Worry"
+            | "X::Worry::P5"
+            | "X::Worry::P5::BackReference"
+            | "X::Worry::P5::LeadingZero"
+            | "X::Worry::P5::Reference"
+    )
+}
